Updated mediawiki/extensions Project: mediawiki/extensions/VisualEditor  ed9477aa5be7a589afd2776bd02b808f786f6141
docsgen: Fix error on Ubuntu

When running this on Mac, it works fine.
But when running it on Ubuntu in labs, the condition
checking for "VE-LOAD" head and body always evaluate to false.

Not sure what's up with that, but since these lines no longer
contain whitespace, a straight comparison is all we need.

Also removed the duplicate `<head>` tag in the template.

Change-Id: Ic7e8351e784be35fd0104269be331cf9bd0c08b4