Updated mediawiki/extensions Project: mediawiki/extensions/Maps  534ccea842f1955e45beb1f74470484ff11346cb null added optional parameter inlineLabel to markers.
This parameter will create floating text right of marker which can contain html
tags <a> and <img>, if the root element is an <a> then when clicked will move to
the page defined by the href property of the <a> tag. if not, then will open
assosciated info window, if there is text assosciated with the marker.

* added javascript libraries needed for this functionality
* added readme to tell where to find these js libraries
* converted indentation to tabs in jquery.googlemap.js

Limitation in mediawiki resource loader prevents me to create a separate
resource module for the added js libraries. I've added a comment on this
in the source code. This has to do with that neither libraries has a
js namespace associated. And resource loader does not provide functionality
to surround a js library with a user-defined namespace. Another solution would
be to manually go into each third party library and change the code so a
namespace is added. However, this is very unmaintanable over time.

Change-Id: Ibe42e589b775856f5f9066f0478a4220a797e5eb