Updated mediawiki/extensions Project: mediawiki/extensions/Maps  03cd4d548ef7ae4265888ab807c5ee55f2e6dbe2 null Made an optional link parameter for markers, lines as an alternative to a text bubble.
This change is intended for the use case where one wishes to create a marker,
line, or polygon to be linked to a wiki page or an external url. and so when
clicking the marker/line/polygon will refer the users browser directly to the
specified url. This was previously solved through creating a text bubble with
a link.

This change however makes it possible to omit that extra step.

example code:
==Marker with absolute link==
{{#display_line:
Bergen~link:http://www.visitbergen.com/
}}

==Marker with wiki link==
{{#display_line:
Bergen~link:Bergen
}}

==Marker with bubble (standard)==
{{#display_line:
Bergen~[[Bergen]]~Is a city in Norway
}}

==Line with absolute link==
{{#display_line:
|centre=Bergen
|lines=Bergen:Kalfaret~link:http://www.visitbergen.com/
}}

==Line with wiki link==
{{#display_line:
|centre=Bergen
|lines=Bergen:Kalfaret~link:Bergen
}}

==Line with bubble (standard)==
{{#display_line:
|centre=Bergen
|lines=Bergen:Kalfaret~[[Bergen and kalfaret]]~Is a busy area of Bergen
}}

==Polygon with absolute link==
{{#display_line:
|centre=Bergen
|polygons=Bergen:Kalfaret:Bergen Bryggen~link:http://www.visitbergen.com/
}}

==Polygon with wiki link==
{{#display_line:
|centre=Bergen
|polygons=Bergen:Kalfaret:Bergen Bryggen~link:Bergen
}}

==Plygon with bubble (standard)==
{{#display_line:
|centre=Bergen
|polygons=Bergen:Kalfaret:Bergen Bryggen~This polygon~Is a busy area of Bergen
}}

This feature has been implemented for google maps and openlayers.

In this change i had to add a Maps_Utils.php file to share identical code
between manipulation classes. But maybe this isn't the way to do it, any
feedback is appreciated.

* fixed indentation (i think)

Change-Id: I269930a19874488a4e93ab3d061fbff7a1a961fb