Updated mediawiki/extensions Project: mediawiki/extensions/Maps  6778773b29f87358e59bd6321b3e71f5c1c99688 null Added circles and rectangles parameter to display_line
Added support for drawing rectangles and circles with all standard metadata parameters as with
lines, markers and polygons. (text bubble, links, stroke and fill styling)

this feature is available with openlayers and googlemaps.

example code:

==Circle with absolute link==
{{#display_line:|centre=Bergen|circles=Bergen:1000~link:http://www.visitbergen.com/}}

==Circle with wiki link==
{{#display_line:|centre=Bergen|circles=Bergen:1000~link:Bergen}}

==Circle with bubble (standard)==
{{#display_line:|centre=Bergen|circles=Bergen:1000~This cicrle~Is a 1000m radius of Bergen}}

==Rectangle with absolute link==
{{#display_line:|centre=Bergen|rectangles=Bergen Kalfaret:Bergen~link:http://www.visitbergen.com/}}

==Rectangle with wiki link==
{{#display_line:|centre=Bergen|rectangles=Bergen Kalfaret:Bergen~link:Bergen}}

==Rectangle with bubble (standard)==
{{#display_line:|centre=Bergen|rectangles=Bergen Kalfaret:Bergen~This rectangle~Is a busy area of Bergen}}

Other code changes:

I restructured some code (in the spirit of DRY), and came up with a solution to centralize certain common
metadata parameters through the use of interfaces. So now, all map elements (marker, lines, ....) can use a
single code library to extract their metadata properties. Basically what it does is that it checks what kind of
model object were working on, if it for example implements iBubble, then we know that this model supports the
parameters title and text, and should be populated as such.

also tried keeping to the code convention of mediawiki when it comes to spacing and indentation.

------

* added base classes
* renamed interfaces to prevent collision

Change-Id: I2599f12a2c7aa81f9c6c52226c787a859949d75b