Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  8a0036db4e16b329eac9ee24d5deb092436025d5 null scrollTo top of page only if no hash present
on slow loading pages if you scroll down when javascript is loaded
the user is forced to the top of the page. This is a terrible experience
we should only hide the address bar at the start of the load.

also make sure do this before initModules - modules like toggle.js
might change the hash and scroll position

Change-Id: I6bc8423693373fb4ddc4e1b7e1b87e5e5b6f230b