Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMaps  b41ef79eebe7b4e8346aae48e1dc4dab11a61cc2 null (Bug 36056) allow $egMapsGoogleMapsZoom to always work
$locationAmount reports *double* the number of locations that are actually
being returned by the query, which prevents the default zoom from being
assigned where only one location is being displayed, since the system
thinks that two are being displayed.

The problem was that $locations already has a value. So when $location is
added to it, we get 2 locations.

The solution was to pass by reference.

Change-Id: Iafec1e1c7092e05e63f9d1cd24c9ee4f94443d5d