Updated mediawiki/extensions Project: mediawiki/extensions/TranslationNotifications  7374bd909a35354eeb13fc2365b54f221c208f67 null Double-checking that the user didn't disable email
If "Enable e-mail from other users" is false, e-mail methods on
Special:TranslatorSignup are disabled. A message in the e-mail part
shows that a user has a confirmed e-mail address, but indicated that
he does not want to receive e-mail from other users.

When actually sending e-mails, the email option is checked again. If
e-mail is disabled, the message is not sent and the e-mail preferences
for TranslatorSignup e-mails is removed.

Patch sets 2 and 3: updated commit message.
Patch set 4: Made saveSettings() unconditional.

Change-Id: Ia1badf6a12adc65263ce8d181dbaed5150fb8593