Updated mediawiki/extensions Project: mediawiki/extensions/Maps  d23dd14a2c6afcf44190c729d459b529d96875a9 null Fixes bug with minzoom and maxzoom parameter
in the case where display_point was used, options.minzoom and options.maxzoom
would always be "undefined", as our if check only regarded the possibility of
value being false or a number, this undefined value caused the javascript to try
to set minzoom and maxzoom when it should not.

Added the proper if checks to prevent this.

Change-Id: I5adf990822afb2597a33f5efda52a7af53a05396