Updated mediawiki/extensions Project: mediawiki/extensions/DismissableSiteNotice  3234c2ec24106c9e7d18d197ecaefbd2f4273966 null (bug 25475) Replacing table with div and fixing directionality.
Applying patch by Nux: Replacing a table with two columns with a <div>
that puts the dismiss link in a float.

In addition, fixing hard-coded right-side alignment of the dismiss link
to alignment according to the directionality of the content language.

Patch set 2: Using alignEnd() instead of a condition.
Patch set 3: According to Brion's comment, made the margin near
             the close link wider.
             Comment: For some reason Gerrit shows changes in the i18n
             file even though I didn't commit them. If I ask for a diff
             between this branch and master, I only see changes in the
             DismissableSiteNotice.php file and not i18n.

Change-Id: I0065abed06180f3cb4b64b81cc35158043d3e8b4