Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  8c796106be9e8f9e8c2fc1fafdb5a50527e9f551 null (bug 35702) Fix link generation on license page
The license page was generating links by the msg() function, which
uses templates and passes in additional parameters. However, when
one of the parameters turns up undefined, the rest are also undefined.
This patch just adds a simple condition check, and fixes the argument
being passed in.

Patchset 2: Use the a element already present, with target=_blank

Change-Id: Ifc0118ef0d20d8ac1aec335a92c53043cc7af7a1