Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  4dbc35d9a80160c9dd39e56a4cddf296c4171de7 null (bug 30225) Add explanation for required field markers
In case it's not perfectly clear, let the user know that a red
asterisk means "required field". Use an existing translation string
(the error a user would see if they *didn't* understand in the first
place) to relay the information, and add some CSS rules to make the
warning appear in the right place.

Change-Id: Ic1b4cdce41686567a96d3ef3d59ccafe2e54d5e5