Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  4a26c0876bd02d0b04f7aca55e8fe4985c4549f9 null (bug 37040) Multi-file select problems should be gone now
Reduce some of the hardcoded browser checks, check for the right
interfaces.

Patchset 2: switch from checking for isSliceAvailable to checking for
FormData interface, with a blacklist for Firefox < 5.0 which has the
bad File.slice().

The non-FormData uploader uses an iframe and cannot process a file
object directly, which is why Firefox 3.6 and Opera were failing
except for the first file.

Change-Id: I2858888fb97f90fa4bd2a159dcc3619669cbd797