Updated mediawiki/extensions Project: mediawiki/extensions/LastModified  a4a58a462c95c3f2136099f45af487137e3f74e0 null Clean-up and consolidation
Making some changes to LastModified so that they don't have to be
overriden by E3Experiments:
  Now handles negative times correctly
  Places message in better location on page
  Doesn't use as many inline styles

Putting all the LastModified JS in a closure and using shorter
function names.

Consolidating a couple E3Experiments functions.

Making code conform to coding conventions:
  Tabs instead of spaces
  CSS class and ids prefixed with mwe for extension code

Change-Id: I5c67d32b75d47ee4ec283cde9653ae4b2556c133

Add nested extension for E3 timestamp experiment

Including this extension here is a kludge, meant to route around a delay
in Gerrit repo creation that is currently blocking the E3 team. The
LastModified extension itself has not been modified and its
functionality has not been impacted.

This experiment is shown to 50% of editors, on a sample factor of 0.006 of
articles. Those in the experiment will see the time the article was last
modified presented in a human-readable format in the article's header.
Clicking on the timestamp will log the event and redirect the user to
the article's history page.

Specific information regarding this experiment can be found on its Meta
page: https://meta.wikimedia.org/wiki/Research:Timestamp_position_modification

Change-Id: Ic57fc3cacbf68d21cb493260c8b566608c681c30