Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  a394e9787607a83785e5af3be3439c4afdb676d6 null Improve description language handling (bugs 31189 and 37293)
(bug 31189) Add languages that exist and have a template
(bug 37293) Add in a descriptionlang parameter, and respect it

This might not be desirable (could use some UX help), but it adds
in any supported languages that have a template associated with
them. For wikis that don't have all the language templates, this
will likely result in fewer supported languages, but for Commons,
this will bring in a lot more options. It's very likely that we'll
need to talk about the implications of that before deploying this
change, but it definitely works now.

OK, there were some tricks I didn't realize existed. And apparently
the Title class was even more complicated than I originally thought
it was. That's annoying.

Also,

This patch will allow passing in a separate parameter that selects
a default language for the description fields in the form, not
necessarily just using the language of the UI. Note that this will
be compatible with the expanded language patch, which is at
https://gerrit.wikimedia.org/r/9528

Now with 75% more mw.Uri!

I also rewrote the library somewhat so it would, rather than
always falling back on 'en', try to fall back on wgContentLanguage
and *then* fall back on the first language in the list (which I've
hard-coded to be 'en'). In future, maybe make this configurable.

Change-Id: Idbac9b6d73870ac85d93ce4efc293886b77a4bce