Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  aa0027e28ecf9be1bfb42b23d306f1476cb7d71f null Rephrasing the frontend-language-header message.
The message frontend-language-header says
"This article is written in $1 languages"

This is a bit confusing, because it may mean that one article is
written in several languages. I propose to change "written" to
"available", but UX experts may propose something even better.

Also adds plural support to the message per bug 37215.

Change-Id: I1417ab4738b48dec7831a811c3c6b02eb71f44c5