Updated mediawiki/extensions Project: mediawiki/extensions/VisualEditor  ee2ddbd3cb060a5473b80015f661df6b9b359ea4 null Fix list handler issues
Lists interrupted by non-empty lines would not close the list properly.
Register for any token instead of just for newlines and close the list if no
listItem follows the newline.

Change-Id: I1743901e3db541bbeda78d17707db943e6ceb9b9

Optionally round-trip sHref in data-mw

If the href would not denormalize, add a copy of the original href in data-mw
and use it to preserve non-conventional capitalization etc.

Change-Id: Ifef50eec7343b0e6b0ba66b6d19a8a3e8c9f8001