Updated mediawiki/extensions Project: mediawiki/extensions/LastModified  4fa65938faa37e4d983c7e67890567be3857b6e9 null (bug 37255) Adjust position of timestamp as needed
The place where we drop in the timestamp may also contain various other
UI elements. We need to inspect the DOM and adjust the position of the
timestamp in light of what we discover.
As an extra precaution, make sure we aren't offsetting an element by
a value greater than its container's width.

Change-Id: I27bf0604b96e972c0a4ff7e8510751496a2f4793