Updated mediawiki/extensions Project: mediawiki/extensions/LastModified  c55526a54af92a94be775fb24bc4c31a64d3a4ef null Emit correct events to ClickTracking extension
I had previously missed on a big part of the spec that stipulates
a particular format for event IDs -- a format that includes the
experiment's name and version and the user's bucket. Additionally,
clicks on the history tab (and not just the timestamp) need to be
tracked as well.

Change-Id: If1b8eab1e9af4358e5d66484436defa93db4159d

Log impressions, not just history clicks

Previously logs were only emitted on the experiment's success condition
(i.e. when someone clicks on the timestamp). This patch adds logging per
experiment impression.

Change-Id: Ic1a541cc2f7833c4da9058ccc9f670372fbe9850