Updated mediawiki/extensions Project: mediawiki/extensions/WebFonts  6e2d7815bb97bf3fa9eb36600ef722e30ca32c62 null Fix Bug 33460 - WebFonts "Select font" portlet menu should only be shown if there are options to choose from
Also address [Bug 32830] Undoing reset doesn't work

If there are not fonts to show in menu, but if there are elements with
webfonts in action(because of their lang attribute) previously we were
showing radio button and a help link.
Now we show a checkbox in that case. Selecting it reset the fonts.
Unselecting it re-applies the fonts for all elements with lang attributes

Change-Id: If2a54f7b4ba8a45eb70d5b92aa44eb5a0b291fd9