Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  85f9c0c3b5e7ae6a9eb7bcceff5cb4294c239729 null (bug 37468) Call the proper error-correcting methods
setError, on the final step, is not a very helpful thing for the
user. They can see the error, sure, but they can't do a thing to
fix it. recoverFromError, at least, will move them back a step.
Along with the fix for bug 32247 (I26bc8538), which adds in a
'remove' button at the describe step, this should solve most
issues that arise from unknown errors on the final step, because
at least you can finish the rest of the files' process, now.

Change-Id: Ie512dac7d86f590cc6ac125b81d1fbcd5f82e356