Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaMaintenance  18873395273043eac998ec9efb5200d3470849ba null jobs-loop: update comment + basic man page
This give a bit of documentation about jobs-loop.sh ala man page. It
might help newcomers figuring it out, or at least give some clue to
people unfamilar with the job queue and trying to fix it on the cluster.

This patch describe jobs-loop.sh as it was intended. Aka we had a way to
pass a job type as an argument to have the loop to only proceed that
kind of type. Some refactoring dropped support for it though, that need
to be fixed in a follow up commit.

No functional change in this patch.

Change-Id: I1633f8d8786b50734e5d1f64985d0cc648415345