Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaMaintenance  913bcb8dec481f5edc1acf37c2e071012e699132 null jobs-loop: fix argument handling
The jobs-loop.sh script used to accept an argument to proceed a specific
type of job. Overtime and refactoring, support for that argument was
dropped since $type variable is expanded by a foreach on $types
basically overriding the argument passed.

This patch restores support for the optional argument and make sure it
will not process any other types (aka the default queue).

Change-Id: I0b9c9eaa7dd03654312d25906f389646a9af5660