Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  49274f9695bd01fc84ae64273dec70869b2661b3 null Fixed formatting bug in table header output.
Before this fix, the following code
{{#ask: [[Category:test]]
|?km2= 10<sup>6</sup> km<sup>2</sup>
|format=table
}}

would produce:
<th ...>
6&lt;/sup&gt;-km&lt;sup&gt;2&lt;/sup&gt;"&gt;
<a ...>
10<sup>6</sup> km<sup>2</sup>
</a>
</th>

where the expected output was:
<th ...>
<a ...>
10<sup>6</sup> km<sup>2</sup>
</a>
</th>

this due the fact that the existing code was trying to
add the column name as a css class to the table header element.
However in this case where the column name contains html tags, then
you would get this unexpected result.

Solved this by adding a strip_tags() on the title before passing it on to the logic
converting the title to a css class name.

Change-Id: Ia70d75f4e53acc68f8c86ba18a6762b0f892fe16