Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  dffad66203b6bc246d8f9ffe1581d28ec3eec08a null (bug 37449) Add in a bunch of license messages for i18n licenses
The default configuration includes cc-by-sa-3.0-nl, for example,
but the default i18n messages do not include those license choices
for people who want to license under those originally (i.e., own-
work deeds). This patch adds in the messages, hopefully correct,
and shoudl fix the problem reported by the bug. However, there may
be a superior way to do this--though I suspect this is much the
same problem raised in the comments of bug 33330, so I may as well
submit that fix in a separate patch.

Change-Id: Ie9f702f4a985433d109d0906f17cfb24c6313e33