Updated mediawiki/extensions Project: mediawiki/extensions/Maps  b986d1224d165cf27880ceeef0ba2134680cca74 null Added an optional visitedicon parameter
This parameter is set as a global parameter as well as a per marker parameter.
When this parameter is set to a valid image url (or wiki file), then when the onclick event
is triggered on a assosciated marker, the marker icon will change to the visitedicon.

A usecase of this is to be able to set a different icon for visited markers, so it is
clear to the user which markers he has opened/clicked and which he has not.

example code:

== Visited icon google maps ==

{{#display_map: 57.70414723434193,88.9453125~Asia; 49.837982453084834,10.1953125~Europa~ ~/static/downloadicon.png~ ~ ~/static/folder-2.png
|visitedicon=/static/tweet.png
|zoom=1
}}

== Visited icon openlayers ==

{{#display_map: 57.70414723434193,88.9453125~Asia; 49.837982453084834,10.1953125~Europa~ ~/static/downloadicon.png~ ~ ~/static/folder-2.png
|visitedicon=/static/tweet.png
|zoom=1
|service=openlayers
}}

Also added the change to the map editor.

This change works with openlayers and googlemaps.

Change-Id: I8b1f7cd52c00585dcb81179be786f7ce05f444e7