Updated mediawiki/extensions Project: mediawiki/extensions/VisualEditor  8d977b236d8a2b38c31403a0a33a82410c993c0b null Use mediawiki.org as the default source in production
Change-Id: I949eebf82ea152b545e6f10764ba4a9865b46144

Revert "Added localhost to available sources and made it default"

We need to keep master deployable, and obviously this won't deploy
right. For now you'll need to keep a local hack in your own Parsoid, and
later we should have a settings file.

This was also breaking people's dev installs because not everyone uses
http://localhost/mediawiki

This reverts commit 0e0d9fbb5060a91dd8ddd343f8ec2eca873d01fd.

Change-Id: Iba69ee62432d8e5f489ee31d9280961a90b79c20