Updated mediawiki/extensions Project: mediawiki/extensions/LdapAuthentication  16f4070c7c340be2621deddc5c3387d68990d019 null Set domain in the user's options
When a long-lived session token is used, set the domain in the user's
options. Add a new function for getting the user's domain, which tries
to pull it from a session, and then from the user's options (if they
have a long-lived session token). Also, change the configuration
settings function to pull the domain from this new function, so that
configuration settings will always be pulled correctly.

Change-Id: If325ea012010e0b3bfaa97bd8572761c8814fcc6