Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  4d6e3a12a345fc26ea479ef3109e08598d085333 null Additional changes to get the normalization right when requesting client side information, and also normalization of UTF8 and squashing whitespace.
Patchset 7: Rebase

Patchset 6: Added a more descriptive error message.

Patchset 5: Shortened the timeout to 10 sec due to lack of maxlag and maxage. This should really have a better solution.

Patchset 4: Commented out use of maxlag, it should be carefully considered if this is what we want in a production environment.

Patchset 3: Rebased after debugging. Bug was unrelated.

Patchset 2: Fixes according to comments, most important is changing Iten
to ItemContent after a rebase, and some whitespace and other cleanup.
There re some additional whitespace cleanup to make sure git sees a changed
file, as it seems like the rebase didn't flag them as changed as it should.

Change-Id: I752625050a6023c4506c0e658fd9081dc2ed32f0