Updated mediawiki/extensions Project: mediawiki/extensions/EtherEditor  a08cfc28a0044228cc97dd3d7914e2ec8364b2c1 null Set text if revision has changed
This patch adds in long-missing functionality--if a page has been
edited since the pad was created, and nobody is currently editing
the pad, we should set the text of the pad to be the current base
revision of the page.

The only thing missing is some way to handle this: A user is
editing the pad, an edit occurs, and then another user comes to
the pad based on the brand new revision. Without that, this is an
incomplete feature, since some people will simply come to the pad
and be confused as to the changes.

My best guess for how to do this currently is similar to the
"fork" functionality, but it may not be the best way. Essentially,
when a new pad needs to be created because of an intersecting
edit, we should fork the pad and give the new editor a PadID that
is not the same as the default. Patch on the way, but this should
go in now.

Change-Id: I12ae725c394af2df45b41eaf6a57c06032c378c9