Updated mediawiki/extensions Project: mediawiki/extensions/Scribunto  be8f6ca34d4de8d4556642a7e2832a7f635c0054 null Allow users to ignore parse errors
Allow users to save a module with a parse error in it, by checking an
"ignore error" box. This was a requested feature in Berlin, there was a
desire to be able to save incomplete drafts of modules. I tested the
effect of #invoking a module with a parse error in it, it shows a normal
script error.

Change-Id: If2978a226ac50150ebeb7837e8dd370a669db771