Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  5d6ddfc582a7edfe5a4efe29a3f54d295d390a5b null Language is sticky now (STTLanguage extension).
This is not yet perfect for wiki pages which have links in wikitext. Right now we use the 'LinkBegin' hook to add the 'uselang' parameter to all link within the page content when they are generated. This will only work when the content is parsed though, so this would require disabling the parser cache to work properly on all pages.

Change-Id: If014728ee209753086707a83793cf80d9cf5df5e