Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  710be9e737f7bc80d15b8b046b8148df529bf515 null Language stickiness now causes cache fragmentation.
A separate cached version will be created when an article is called in a specific language by using 'uselang' or through the defined user language. This fixes the problem that the links in the article were always pointing to pages in the language of the user who had triggered the parsing of the page.

patchset 2: added some more documentation

Change-Id: Id6c35476d86831a0792110b9c8493f7c7c7f7db8