Updated mediawiki/extensions Project: mediawiki/extensions/EtherEditor  7f9dbb6808fbec1b9c38c9aa543a6facd6cc4bf8 null Make tests, make the tests work.
This is the first stab at making a testing suite for EtherEditor.
I could decidedly use some review on this one, to make sure that I
did it all right.

I only added tests for the API calls to start, but I'm definitely
interested in testing the JavaScript side of things at least a
bit, and maybe even trying to test the other PHP in this
extension.

Change-Id: Id34838f89126178cf2d6657531f2606d7d050eff