Updated mediawiki/extensions Project: mediawiki/extensions/EtherEditor  cb367a6a5c784baebe6081ddba4b4dcda8dc2a98 null Add tests for EtherEditorPad, and make them work.
This is a further addition of phpunit tests, so now the entire
codebase should be covered. And indeed, I have the code coverage
files that prove it.

Well, the last remaining things to test would be the remainder of
the frontend and the actual hooks themselves. I don't currently
intend to do that, though, so it looks like I'm done with unit
testing for now!

Patchset 2: Make most of hashar's changes, at least some of which
            were incorporated in the commit we depend on.

Conflicts:
	EtherEditorHooks.php (one include addition)

Change-Id: I8cc6465c73010c32b80f18c4ae0350c233620f2e