Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  5026e3208daa9917954c69ba82ce4f2387f687b0 null Core tests should all run again now.
This is achieved by disabling certain functionality of the STTLanguage extension when tests are running. This is necessary because MW parser tests don't consider the different outcome certain functions can have since they are offering the hooks to change the output. In this case the modification of all URLs is a problem the MW core tests do not expect.

Change-Id: I847554a0cdedee83a9d8cea69b9e059026be7309