Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  051297f95d633171752709deecd118b8a0655bc6 null Behavior fix in site suggestion for site links
As Denny requested, I changed the behavior of the sitelink interface so the first suggested entry from the suggestions list is chosen as the site when leaving the field.
Some minor, related cleanup was done as well.

patchset 2: (tgritschacher) changed the selenium test to check this
functionality too; also simplified sitelink tests a bit;

Change-Id: Ic8e7d85365d8facf0b41233295feaf7755523c51