Updated mediawiki/extensions Project: mediawiki/extensions/EtherEditor  dbbc4b70899d562dd5f6741ad0856a05610f514e null Make the UI more integrated with other editors
This is a big patch (it's 0.3.0!), it moves all of the UI elements
into whatever editor is available at the time. This means we no
longer use the ep_wikitext plugin, now we just use ep_ethereditor.
That plugin is a lot smaller, among other things.

In any case, this also means that there's no longer a separate
page for collaborative editing.

UI fixes will follow shortly, here, once Munaf and I figure it all
out.

Change-Id: Ib25a78a22849fbb3815bacd4031e87a5a837bb1b