Updated mediawiki/extensions Project: mediawiki/extensions/ArticleFeedbackv5  fc3ace8ef2ed1b8aa5cc3870f5cc3de834f6b8b6 null fix DB install when ArticleFeedback is already there
The ArticleFeedbackv5.sql would never be applied if ArticleFeedback was
already there. The check for freshness was done on the old
article_feedback table instead of aft_article_feedback.

Change-Id: I7cea6b72e50563e387496457752ec0e15c5eaef8

fix install on SQLite

The `aft_article_answer_text` table definition was incorrect for SQLite
processing which expect 'PRIMARY KEY' to be passed along AUTO_INCREMENT.

Untested on MySQL

Change-Id: I365ffb218653d1c6d2f287021d1ac8c99e3bcb84