Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  ebc06e4b0c20e7804cbb41c0aa5f169182ab7231 null Make the language fallback code more solid and reusable
In some weird cases the internal structure can be erroneous, or
the language code non-supported. In those cases it is necessary
with a lot of additional tests. This patchset will do some of
them, but also relie on the Language class to behave, which it
does not. At some point the caching in the Language class should
be fixed, but that is a core issue and not part of this patchset.

Change-Id: I6598d400d4866720789d9ebba62d523bc8e4a9e4