Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  5e5154d97ddb07db5859f4e45079f0636211676b null Use only the list of supported wikis
Previously the list of all valid wikis was returned, with this
change only the list of wikis belonging to the Wikipedia group
is listed as valid. The members of the group is as it is defined
at any time in the internal table, or its fake alternative, and
may or may not have a direct resemblance with sites listed in
the sitematrix.

Patchset 2: Moved getGroup outside loop, its more efficient that
way compared to getting the group for each sitelink.

Change-Id: I4d672067da502cc3fe497db2335ef5622b2f9f9d