Updated mediawiki/extensions Project: mediawiki/extensions/SimpleFarm  4a29df5f6efac84fed0380f56fb2692934832a9c null Minor Todo added
Change-Id: I25c44e116ec6b7b1dd423795fbe4f59d8c3292c8

Splitted includes and moved them into includes directory

Change-Id: I8808ff4784d6697c45e17fc0994e387dfd2d02f4

Less trickery when detecting farm members using the path option.

Now, the actual original path is simply compared with each members path and if the members path is part of it, the member is qualified. The more specific a members path is, the more qualified the member is.

Change-Id: I807a3fc79750e8dcc42cfab598ea383e498cd082

Spelling fixed in some of the code docs

Change-Id: I3837360a2f5633b322066bcd66d877915adda428