Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaIncubator  26c18bc2edfec2ab621a6405f83fd5b6705ebfab null Make separate functions for better code readability
Call analyzePrefix with a title objects which checks namespaces.
Make a separate function for the second part of onShowMissingArticle. So the biggest part of the commit is just decreasing the indenting.
In general, instead of "if( no info page ) { do 1; return; } do 2;" it is now "if( info page ) { call other function for 2; return; } do 1"

Change-Id: I8367f99a8e23eda4034d6e1b51afc477cb284f92