Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  b86d881c06ecd7b057ce1e4a7ada59136fb20f14 null Changed stashed upload success message to be clearer
The message used to say just "OK", which is kinda clear in context,
but can be clearer. It is also very confusing to translators.
There was even a comment nearby, saying
"this is just "OK", say something more."

I changed it to "Uploaded", updated the qqq documentation and
removed the comment.

Change-Id: I3953f49912681948789dcbbe2370c131e1ad07d1