Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  91b1d987fc85e04d1ef2e265ca3926d8bf843282 null Switching priorities to text labels and a bunch of style fixes
Change-Id: Ia46e62c8fa6e34d91b74e5c0722714fdc25dc5df

[FR 2012 Card 6] CentralNotice: Add Z Indexes to Campaigns

Fixing a bug where a preferred campaign will override, and display nothing,
a campaign that does have content.

It does this by introducing the concept of Z indexes which will allow
campaigns to be stacked on top of one another. If a high priority
campaign has no content it will fall back to a lower one which does.

It also now normalizes banners in the user's browser. All campaigns
on the same level are treated at the same priority. And all banners
in a campaign are now relatively weighted after filtering.

The allocation tab now reflects this new allocation strategy as well.

Patchset6: Fixing bug in chooseBanner

Change-Id: I94537d2971c333eb175675a4607454fe3e4700fa