Updated mediawiki/extensions Project: mediawiki/extensions/E3Experiments  4d36b4ab6f8874cb5ae2c45c572f1e57608e0098 null track() returns a promise
HTTP 204 responses (which is what //bits.wikimedia.org/event.gif has
been configured to return) reliably cause the HTMLImageElement's
"onerror" event to fire, so although it is semantically confusing we
can use it to mark a successful tracking call.

This isn't fool-proof -- DNS resolution errors will also fire
"onerror", for example. But this is OK because track() is really
a one-way, fire-and-forget way of pushing data to the server. We don't
want to get entangled in things like retrying failed requests, etc. So
the promise's resolution means "this request is as done as it's ever
going to be", which is what we want.

(Krinkle, I meant to include you on the original review request. If
 you have a chance, could you look over the whole thing and not just
 the diff?)

Change-Id: I7c05985abd3b658f21c81d32cdea557b75407a08
References:
 * http://www.phpied.com/ajax-with-images/
 * http://goo.gl/4AkF4
 * http://javascriptturnsmeon.com/image-beacons