Updated mediawiki/extensions Project: mediawiki/extensions/E3Experiments  4c16d155cd4b748637ee7669c0d0003f23be8bbb null Add dummy tracking call from Special:RecentPages
The Varnishes on bits now expose a URL that returns HTTP 204, to be used as an
endpoint for pushing event data to our servers. There is no data collection
being performed yet, but we still want to know for certain that handling these
requests doesn't cause Varnish to choke. This change adds an event call from
Special:RecentChanges, which gets about 700k hits a day, making it a good
candidate for us.

Change-Id: Ic3a5a114bddc62cce42fa292bc086687eacb197a