Updated mediawiki/extensions Project: mediawiki/extensions/SemanticResultFormats  7b10f79c0e24aec1f39f5aadfb55424a53449deb null jqPlot re-factoring Part II (relates to bug 35627) (major)
## Description
Part II of the re-factoring effort was to ensure that the jqplot printer
and its JavaScript plug-ins are used  throughout RL, eliminate identical
code among jqplotbar and jqplotpie classes, and make use of the
same JavaScript for both jqplot.series and jqplot.chart class.

The consequence is that jqplotbar and jqplotpie merged into the
jqplotchart class.

Additional message items will be posted in a subsequent commit to
avoid conflicts with l18n.

A more comprehensive description can be found under [1] and [2]

## Notes
[1] http://www.semantic-mediawiki.org/wiki/Help:Jqplotseries_format
[2] http://www.semantic-mediawiki.org/wiki/Help:Jqplotchart_format

## Patchsets
+1 initial commit, added test files
+2 kill whitespaces
+3 reinstated the overriden $moduleTemplate definitions
+4 rearranged and merged js files (avoid dublicate code saving
15% of code), jshint checked (all its conditions are met)
+5 kill whitespaces

Change-Id: I3c8847aa9072f7302346f0b272e3bba15cac3298