Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  95368241fd490ac771c636a8b7e624f6fefebb38 null Added filtering of aliases on languages in getAllAliases
This function has no explicit test in EntityObjectTest, but is
used in several parts of the overall code where correct behavior
depends on this function so it already have tests in place. For
example ApiGetItems.php line 116.

Note that there is a default function argument so the function
will continue to work as expected.

Change-Id: I500eea5583162a695e3172a0a1dbe6029256929f