Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  414587aa100ea22a8d56bd4bf271e03dcccd7cf4 null (bug 36791) replace date fields with jquery.ui.datepicker
The tricky part of this fix is the datepicker "altField".  In order
for datepicker to internationalize the dateFormat, we add a hidden
input with the date in a known format, and shadow the user-facing
date.

I didn't touch the bug in SpecialCentralNoticeLogs end date,
where an end date filter is interpreted as midnight of the
previous day, so for example a filter of 8/13 - 8/13 will
never return log entries.

There might be problems using strtotime() on old versions of PHP,
please report if encountered.  The symptom would be that newly
added campaigns have a duration != 1 month.

https://mingle.corp.wikimedia.org/projects/fundraiser_2012/cards/439

Patchset 12: Using ISO 8601 for user-facing date display

Change-Id: Id708c8d3a512513d98442127a11f02bb06039d97