Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  b795a5b2ed826d3b6e7feff39695bf7ac9c8a144 null Items in non-editable views are not editable anymore
Before it was possible to edit items for example within a diff view or an old revision view of the item. In these cases edit toolbars are now hidden.

Change-Id: I8489d8bb831ca4c9dfea842215a1edac9c9dd693

Tell JS whether editing should be enabled.

This is done via a JS config variable and a marker element in HTML.

patchset 4: Fix permission checks for non-existing items.

Change-Id: I633fb8303aa1bf9b6432431d8736caf826d97241