Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  aed7add2bf08518a7b1dc99a5e69ee9e3ed6c2be null Removed url args for usekeys/nousekeys, and also setting through config
This is now completely under control from the format, and only one form
is returned and that will be with keys unless it is not compatible with
the format.

There are tests to check if the keys in the format is according to some
of the actual values.

Patchset: Rebase...

Change-Id: I4eaa36de9c85c7de590092bd4d0fb6e4b51d7911