Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaIncubator  c47b75bc64e30b8c31176a1ed6e4387fe07327cc null Rewrite logo/configuration code
Rewrite and improve code related to site configuration and things that depend on it (especially the logo).
There is now a general getConf function to get a specific configuration, and use it for getSubdomain, fnTestWikiLogo and InfoPage::makeLogo

This means that the logo on the info page now uses the localised version as defined in the configuration, and that the fnTestWikiLogo feature should now actually work.

Change-Id: I65b68eb8182d94f8b169038ec99ad584436a238a