Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  f31fd238b6e1100a188c114db0e28c4f1ca7fbcd null Handling of "gettoken" in wbgetitems and wbmodifyitem
This only handles inclusion, and exclusion, of tokens. This solution
still uses edittoken and not the correct itemtoken.

See https://bugzilla.wikimedia.org/show_bug.cgi?id=38521

Patchset 2: Removed some stuff not used anymore.

Note that this changes could later be removed if the
preferred solution is to always use a single token,
that is the edit token.

Change-Id: I7c72f6e2c21e6835e63ac92fe3e021d06b1a7af0