Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  f9c076c8ad8508e26e42fe78ac74dcff15d8692b null (FR #595) Adyen payment gateway
Implement credit and debit cards, using an iframe.

Introduce a new technique for creating an iframe: javascript
calls a php api which builds staged variables, and this is posted
to a form with target= iframe id.

TODO: It should be possible to use "communication type" to
make this a bit more consistent with other adapters.

Note that the files in forms/ are mostly dumb copies and need
hacking.  They were committed to illustrate that adyen.js has
some new code to support the above iframe technique.

Send the custom filters risk_score to Adyen as an initial fraud
offset.  Add features to base gateway adapter to provide access
to the pre-processing fraud score.

Fail if preprocess filters indicate fraud.

Mark transactions returned as pending and send to the -pending queue

Fixing double staging issue in both GC and Adyen where we could
potentially make the charged amount larger. (Multiple x100s!)

Change-Id: I049d245c37cda0c0799cd82dda674595230cae07