Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  7411251c7511ebe038d1f062839a1872790fa97b null Improve robustnes of URI values against errors
The old code tried to create an internal dataitem even in case
of errors, to ensure that there is always an object to access.
Some cases were forgotten. For a cleaner solution, the new
code creates a dummy object on the fly if needed for some
operation. Should fix the reopened bug 30717.

Change-Id: Id5e48f16e769e97b0a5aece85664c2c031a6ea12