Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  57990f83cac51c6618878443d1ed5bf5992c40cd null Fix error when calling EntityContent::save repeatedly.
EntityContent::save() would save if called repeatedly on the same
EntityContent object, because the entity's WikiPage object would
detect that the new content is the same as the old content. Which
is the case only because the old content object was actually
modified.

This is fixed by calling WikiPage::clear in EntityContent::save.

Change-Id: I5c5330a921d27e9bed093ed8cd98c40db7988895