Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  cec56cce659397df5f239e394f6eeb3455448eb8 null Improved editable aliases and prepared its DOM structure for grid layout
the AliasesEditTool doesn't initialize with an empty EditableAliases anymore. Instead we initialize an empty AliasesEditTool if no aliases are set yet, and initialize the EditableAliases instance on demand. This solves problems we had before with the button and solves some minor hackery.
Also prepared DOM structure for grid layout. This requires that aliases label (part of AliasesEditTool) and the value (part of EditableAliases) are structured within one DOM node to have them together as block element which we can give some padding on the right to overlay the edit tools.

- patch-set 2: added QUnit tests for AliasesEditTool and implemented destroy function
- patch-set 5: rebase and fixed conflicts
- patch-set 6: adjusted tests for the new aliases DOM structure

Change-Id: Ie07985986450550d1efa9a27770b2d62bfc01e41