Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  23a694e7b9e893616257346c957109700c1557fe null (Bug 39485) Encode params for ItemByLabel and ItemByTitle
The two pages are using messages that constructs url links.
Those links should either url encode their argument or the code
should provide the param as a pre-encoded argument. The fix
provides both forms, where the first one is a url encoded variant
and the second is raw. The message is fully parsed so the text
will be properly escaped.

Change-Id: I36e6abf53012232e39091954a53af4525520341e