Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  01fb7f3b6717ee3570f7c452eec572396b184465 null (bug 39619) Be more skin-agnostic with the scrolling commands.
There may be further skin-specific things in this extension, but
in the interest of getting this bug fixed and pushed out, this is
the only necessary change.

Basically, the firstHeading ID only got added in certain skins,
and the other skins were left out in the cold. Now, we do this
properly.

Change-Id: I29c71fb48ea1a83f398ea6d3ebed1cb7be91b3f2