Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  d4e24d023fc496e21bef58ea61ba7389fbe552ea null (Bug 39500) Get rid of getUsekeys settings
This does not use exactly as the bug says, as the getUsekeys as a
method is still there, it just does not get its settings from the
config anymore. It still has to be defined somewhere and it is
now hardcoded in an array inside a static method Api::usekeys().

I really don't like this because it locks the code to the existing
formats. If someone makes a specialized format and wants to use this
extension, then he must patch the extension to make it work with the
new format if that does not support use of keys.

The present formats that does not support usekeys are XML and WDDX
(WDDX is encoded as XML) and their pretty formatted versions.

Change-Id: I10f35bd61fdba239d5e4c52f51b5466dcbd33c7c