Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  24349c4b142c0a3a41cfa63df8e5b7c199adb5f2 null (bug 39608) Check for undefined in many places
O_O

Several bugs converged:

1. In jQuery, $.each doesn't check whether an array member exists
   or not. So var a = []; a[1000] = 10; $.each(a, fn); means fn
   will be called 1001 times. Just silly.

2. In UploadWizard, we don't splice the uploads list like sane
   people, we just delete items from the list.

3. In UploadWizard, we assume that all of the uploads are UWU
   objects, and use $.each to traverse the list.

So now, I added a bunch of instances of if ( upload === undefined
 ) return; and threw in some checks for the upload counts. This
bug is fixed now.

I think we need to revisit this, at a later time, as a symptom of
us doing things sooooo wrong. We could just splice items out, but
it will require some other changes, so I won't deal with it now.

Change-Id: I4994d2bf7c75845ac31305cf553a1c5ce1526bc0