Updated mediawiki/extensions Project: mediawiki/extensions/ArticleFeedbackv5  139578dfd707f6b8d9b46a1e19da70ca2ae1d277 null Fix flyover positioning for RTL environments
Before this fix the flyover would appear on the wrong side in
an RTL languagel.

Positioning the flyover itself is easy - just matching the gravity
to the <body> directionality.

Making the arrow look good was hard, however, and can probably be
done with much cleaner CSS.

At the time of submission this patch works well on the master branch
of AFTv5 and with core fix I5db952bc applied.

Change-Id: I0d153c5860e6c364a4f3319531c0b8540e8a6bb2