Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  aae6edc534593f79aa7e06726eeb86d0f0e494b7 null Fixed bug where aliases would be broader than grid layout
The max-width for aliases in edit mode was set to 1000. This is now flexible and set to the current width of the aliases <ul> list. We also do this max-width calculation for the helper tag provided by the tag-a-data widget, this was not done before but nobody noticed because the default here was 1000 as well.

Patch Set 2: will also work for the helper tag now.

Change-Id: Ic023f5662ae98dbeb3835392194f2d0887067c7c