Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  4b2c1a65cc157bde8345732c4acbed9d9cae0ee9 null (Bug 38843) Refine behavior of wbsetitem
There are a few remaining issues with wbsetitem, the most important
is that the input format should support what we call "the long format"
which is as close as possible to the output json format.

Refer to the documentation for wbsetitem for a complete description.

Patchset 3: Cleanup of duplicate code

Patchset 4: Rebase, fixed exceptions, uncommented a test.

Change-Id: I8d833861f7a4540b0c03a11b936be6bd9f6c0219