Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  7744e136e963a90f44149abc3fa722f38c09592d null Don't allow setting of internal IDs.
Internal IDs should never change, and they are determined by the database.
They should never be set prograsmatically.

This is especially true for the test sites:

Setting the internal ID directly causes ORMRow::hasIdField() to return true, so
ORMRow::save() will try *update* the database row instead of inserting it
(which for some reason is still reported as success, even though it
does nothing). This causes TestSites::insertIntoDb() to leave the table empty.

Change-Id: Ifcbb5a9be0f6603cd46de1bcd469997e9a93f908