Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  30153bbd2da9a69a0a370fa153634949b3e4e52b null (Bug 39839) Add a faster method for "userWasLastToEdit"
The present method loads a number of revisions and then loops over
them while trying to find a conflicting user. This is somewhat slow,
and by adding the check in the database query itself less data is
transferred.

The present method also have problems after a merge of history
(now it only have a small problem withing a sub-second), and it have
problems with anonymous users as they share a common id (now it
only have problems with anonymous editors with a common IP address).

Patchset 2: Rebase

Patchset 3: Fix according to comments

Change-Id: I03e0b39e83491b35e863a1a94fefbc25e27791d3