Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  66a30b0dbaf68c351ec71f8a9c161cc8edd25e9b null (Bug 37750) Sort sitelinks before listing throug wbgetitems
As a simple fix for "New added interwikis are appended
instead of sorted into the list" the list of sitelinks
is sorted before being used for building the list in
wbsetitem. Sorting is according to the global id and will
give a fixed sort order, but not one that all users
are comfortable with.

See also change I7685e904

Change-Id: I4ffd8be719d295290cacc92c6cdd37c5d6d10097