Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  ec73842c4485060009d2f468baa86df23ec33083 null Added wb.utilities.newExtension, solved some minor bug in Tooltip.ext
New utility function for declaring an 'extension' which can then easily be used to declare generic pieces of code which can then be extended in more specific implementations.
The old Toolbar.ext serves as an example how to use the function, at the same time as bug was solved which would break tooltip behavior when using setTooltip() with a string provided.
This could be further extended to be more useful, for now this simply serves as convenience and should help avoiding bugs as the one of the previous 'Tooltip.ext'.

Change-Id: I7fbe0a4c65347fd8cdb316261adf66da3e771bf7