Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  a84c69c0cac9b6e41e9f2999ad43106cc6de3985 null use localstorage instead of cookies for dismissing banners where supported
this uses local storage which is preferable as it doesn't cause cache
fragmentation

Change-Id: Ibb76cfc20f8a7b44e8399b559004090cfe54e3e2

hide banners targetted for android only

Change-Id: Ie8ae7d65b6d2d0402e90d93b72b8a1cce2872e01

hack to limit cookie fallback to only zero rated banners

setting cookies introduces cache fragmentation. This limits the setting
of cookies only to zero rated banners on devices that do not support local
storage. Ideally we should fix the vary headers to allow setting of any cookies
but this will have to do in the meantime.

Change-Id: I10fd5b1c2d044ba4643c1d78c0100ed64ae2d795

refactor banner code to be generic

tested on ZeroRatedMobileAccess
added stub to SkinMobile.php looking for backend to suggest best way of
serving custom html. Ideally this should be configurable

Change-Id: I1e36c3a658d38646964c3887c19d3a82c3e63ad7