Updated mediawiki/extensions Project: mediawiki/extensions/Translate  e1699acf2f120694a708e777c5f28b711433c476 null Add extra = null; to MessageGroups::clearCache .
The lack of this has been the cause of a couple of problems
for me, because it gets cached early and then sticks, causing
problems when trying to add/remove groups on an ad hoc basis.

Groups added on an ad hoc basis are in use in TranslateSvg
and are required for good unit test coverage.

Change-Id: I534120053ab152e821feedc1dec3f07636d914b3