Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  0aa5d843647a0d97bdee43e93c3445175b836a53 null Added tests for ItemView
ItemViewTest now also checks for the itemId related class/id values on the outer dom node.
Changed EntityObject and all inheriting classes to have a public/static getIdPrefix() instead of a protected one since there is nothing to hide about this and this can also be valuable information in a static context.

Change-Id: Ie84ec8bea69382aa9ea98d7000bf29f5e387e8b4