Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  98d468d2cf65e1ccb5461eae19cb0adb91e03682 null Prevent direct editing of structured content.
This patch prevents the use of the API's edit action on
any content in the entity namespace to avoid text-level manipulation
of the JSON content. This also prevents content of the wrong type
in entity namespaces.

Once I37b0e42e is merged, this patch also prevents entities
to be saved to other namespaces than their own.

Change-Id: I4d554f4e35ef4a5fa3bb2051ea0001480ace0ca2