Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  bc3723ef44d966997eeb5e53d7a691988ab1e622 null Added some JavaScript to allow quick navigation through long diffs
Making subbu's life easier with accesskey and jQuery :)

Use shift+alt+n to go to the next diff, and shift+alt+p to go to
the previous one.

The jQuery library needed to be included, so we have that in our
repo now. I could change it to the minified version if necessary.

Change-Id: Id1226df3378428b79d684d02f2d40669e4aefe10