Updated mediawiki/extensions Project: mediawiki/extensions/CheckUser  ede4937ffbd887530c01785a95be349236990ecb null Store private data encrypted with a public key
Keep a copy of private data encrypted with a public key, so that it
is possible to recover the data by using the corresponding private
key.

Adds class CheckUserEncryptedData, which is meant to be instantiated,
serialzed, and stored in a database. Later, they can be unserialized,
and getPlaintext() can be called with the appropriate private key to
read back the data.

patchset2: Cleaned up whitespace / debugging

Change-Id: I0044e38f5fc9405fbcad097c2f6c30719af97b52