Updated mediawiki/extensions Project: mediawiki/extensions/WebFonts  1cd0b7552b9e858942f5a944e86c5a9457b4eb0e
(bug 39381) Add Javanese font to WebFonts

Mapping the font to jv-java - Javanese with Java script.
That meeans, this font will not work with the usual user interface
or content language trick. Lating is the default script for jv.
But still font can be used by explicit tagging of the content
with lang=jv-java attribute.

Change-Id: I0fbae92178d9f6bcfa0a163375bc1cbe2fec6876