Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  d17bf4f1091f873847b0beb68a01c78c45fde43f null Comment out alternative configuration so people can still include the examplesettings file
Change-Id: I0ca24db5edaee09c02a351f730b422a21e4eabff

Fix change notifications for deletions.

Directly accessing the text table is not going to work.

Note: this needs Ie1fc967d on the Wikidata branch!

Change-Id: Ia5bf996a9b2cffff242133e391de16ef1a549e15

[Bug 40390][Bug 40397] entity namespace fix.

This change makes the lists of entity namespaces and entity models available
programatically, and uses these lists instead of namespace constants everywhere.
It also fixes the behaviour of Wikibase.i18n.namespaces.php to not interfere
with core namespaces.

WARNING: this change requires a change in configuration of all existing installs!
Without the necessary configuration change, old installs will abort with an error.
The ExampleSettings.php file has been updated to reflect the new settings.

When merging, please send a mail to the list alerting people of this change!

Change-Id: I3e96a402fd4fa9f70c737abc3fb7b37c6b27cb35