Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  e12e1e8579c08fb6ca505097f043df9830511067 null (bug 40399) site-link urls will be taken from API result
This is achieved by some hackery where we overwrite the wb.Site.getUrlTo function to add a filter for all urls we are sure about because we got them via the API. This seems like the easiest way to implement this without refactoring more. Also, it still allows to use wb.Site.getUrlTo() without the API, which should work rather well (at least for MediaWiki sites) now.

Change-Id: I23d54b0969073dc0188095b0405d7a78b931e200