Updated mediawiki/extensions Project: mediawiki/extensions/skins  6d4bf8cdfe978f4ca406e9de91289139e506bc4a null Tomas skin: Fix various issues
Fixed a reference error in Tomas where it assumed extensions
would always be in wgScriptPath/extensions, and even worse, where
it assumed that itself is in $IP/extensions/skins.

Removed wgSkipSkins configuration. If that is wanted, it should
be done from LocalSettings.

Fixed wgValidSkinNames. Keys are expected to be lowercase. The
value is the human-readable interface name (e.g. for Special:Preferences).

Change-Id: I1bccddff080537c907a9b2bea21dbef2a7c7558e