Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  9ea500ca48a95906c15808ce58c360e088bc4354 null Fix regression introduced in 992a87 handling template-affected attrs.
For this example below:

<div {{echo|style}}={{echo|'color:red;'}}>a</div>

the regressed code was not marking the value as being template-affected
whereas it actually is.  The regression was caused by an incorrect check
that detected if the key was a html-attribute (mw:content is not a
html attribute, for ex.  It is a parsoid-internal token attribute).

Change-Id: Ic4c327fd130614b98a268d82d081f73801ca4d2d