Updated mediawiki/extensions Project: mediawiki/extensions/BreadCrumbs  58fa042e8205d01fdeff637af4faf122b6a9837f null Fixed Session bug, attempting to follow Style Guidelines
This commit fixes a number of small issues:
1. BreadCrumbs were not being written to $_SESSION in some wikis
for anonymous users. The problem was that this is the expected
behavior for wikis.
The fix was to force the wiki to set up a session if one did not
yet exist, assurring that the content would be saved across page
loads.
Discussion: https://www.mediawiki.org/wiki/Extension_talk:BreadCrumbs#Breadcrumbs_only_showing_current_page_if_not_logged_in_18230

2. My Company's use called for the ability to prevent recording
breadcrumbs in the Special Namespace. I added generic namespace
exceptions which can be specified by admins by pushing the
namespace's title string to the array $wgBreadCrumbsIgnoreNameSpaces.

3. The Hook BreadCrumbs was using only wrote BreadCrumbs to
user-editable pages, though they should be available on Special
Pages as well. This is now fixed.

4. I'm trying to adhere to Mediawiki's style standards, both
for coding style and version handling, though the latter is
fairly unintuitive. I restored a translation that was
accidentally overwritten in a prior commit, revised the array
orientation in the code, updated the message handlers not to
use the global messaging function explicitly, per the
suggestions of Sumanah and kghbln.  Thanks a lot, guys!
ALSO, a further change was ammended to correct the trailing
whitespace in my PHP.

Original Commit Message:
Overwriting Commits on top of Abandoned Dependency.

Change-Id: If0d10aeb479cfb9f8878e05a39f1bd1c4fa504b4