Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  34803958a6b28712aaec101f7915f6397d49970a null Improved wikibaseAutocomplete behavior, don't move cursor unnecessarily
In the case the the suggested autocomplete value equals the current value, the cursor was moved by the autocompleteString function. This is annoying in some situations, e.g. Typing "Urk" which is completed to "Ukraina", then hitting ctrl+A to select everything (perhaps one wants to copy/paste) but in that moment autocompleteString comes and sets the cursor to the end of the input field.
Also changed this to use third parameter (with "backward") in setSelectionRange function so one can just press arrow left to loose the selection without suddenly having the cursor at the end of the selection

Change-Id: Iaebb34e566cadb2dbe9eaf936ad03d431464162d