Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  d4c74ef1b3a7d7a77f411327b3adef22e1205912 null (FR #566) multiple sets of account credentials
Each set has a name, which is a key into $wg{prefix}AccountInfo.
Note that this fix requires changes to LocalSettings... as the
code is currently implemented, the only safe migration is to
create the new account info variable in parallel to the existing
globals.

A new field 'gateway_account' is added to queue messages, and
contains the account name.

TODO:

* Logic to choose an account based on the incoming params is stubbed.
We always use the first account.

* Gateways still in code review must be rebased and updated.

Change-Id: I5b88f15e6c6255c9b7f44ad770f36f47a58d8d40