Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  ce631c2238af1139428e5e227d2daf4b2d2bd562 null Addresses bug #38009
* Makes hostname for stopMobileRedierct cookie configurable
If wgMobileFrontendStopMobileRedirectHost is not set, MobileFrontend
will automatically set it to the base domain. This is necessary to
easily expose the correct domain to javascript so we can unset the
stopMobileRedirect cookie client-side.
* Makes localBasePath and remoteExtPath variables
This reduces repetition and will makes resource paths easier to
change in bulk if necessary.
* Adds a JS resource to desktop site for unsetting stopMobileRedirect cookie
* Adds wgCookiePath and wgMobileFrontendStopRedirectCookieHost to
mw.config vars
* Adds mechanism to load desktop-specific resources in
ExtMobileFrontend::requestContextCreateSkin()
* Wraps desktop resource output in conditional dependent on config setting

Change-Id: Ic21e3dbc070c8bad3a10f5cd17f5fef4c3dcf90a