Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  7739922f5d61ef9bea2cc21abdaf9d285f6c5c0d null (bug 37361) (SMW Roadmap) Info tooltip replacement
## Description
This commit contains all relevant methods to replace the current SMW
info tooltip implementation. The former implementation (SMW_tooltip.js)
is still available for legacy purpose.

## Implementation
First we have tried to use MW's tipsy plugin but it has shown to be inflexible
for service links that expect a tooltip to be open until an event is
triggered therefore the decision was to use qtip2[2](v2.0.0 - 2012-09-10)
as third-party juery plug-in instead.

## Tooltip
The tooltip [3] will act in a hover like manner for entities marked with
smwttinline (properties etc.) and a persitent tooltip marked with
smwttpersist that allows interactions within the tooltip. The persitent tooltip wiil
be open until the user triggers an event that hides the tooltip or use
the close button etc.

## Configuration
+ Objects marked with smwttinline will adopt a hover like behaviour with no
close button and automatically hide by crossing the tooltip border
+ Warning tooltip will behave such as smwttinline
+ Service link tooltip will have a close button and stay open until a event
is triggered
+ Info tooltip will behave such as a service link tooltip

## Tooltip styling
See [4]

## Note
+ Messages added
+ jshint checked
+ Using CSS background-image instead of <img/href> generated icons
+ Resource is loaded at the top otherwise the stylesheet will only
become active after all content is loaded with icons appearing with a
delay due to missing stylesheet definitions at the time of the display

## Solves
(bug 29764), allows the size to be increased automatically
(bug 39445), corrects distorted markup in service links

## Should solve
(bug 33443), problem with legacy js code
(bug 22951), N/A
(bug 40644), allows to specify a service icon via CSS to adopt
individual requirements

## Links
[1] http://semantic-mediawiki.org/wiki/Roadmap#Replace_.23info_pop-up_code

[2] http://craigsworks.com/projects/qtip2/ is free under the MIT/GPLv2
licenses and supports C8+, FF3+, IE6+, O9+, S2+/iOS4+

[3] http://semantic-mediawiki.org/wiki/File:Smw-1.8-tooltip.png

[4] http://craigsworks.com/projects/qtip2/docs/style/ and
http://craigsworks.com/projects/qtip2/demos/tips

Change-Id: Ida6ebae42594b0017902dc4394a83caa63b55368