Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  ceaeae75476a5e6f853542556ec8b39e70e9ddba null CN/TrEx 3 -- Message Group Review & CN Namespace
Similar to the abandoned https://gerrit.wikimedia.org/r/#/c/24638/
which had to be abandoned because git review is dumb...

Using the message group created in CN/TrEx 2; we now apply translation
review features to CentralNotice messages. This is done by creating
messages in a new CentralNotice namespace (allowing everyone to edit
banner text) and then moving the messages to the MW namespace when
the review is changed to the correct state.

States are defined in https://gerrit.wikimedia.org/r/#/c/26769/

Change-Id: I73c5716807567aecce19584bf34744f5097edf1b