Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  acc96729d21f69d1144df9914162a0ce5ac1f8b3 null Fix for failing QUnit test for EditableValue
This test was actually wrong before. The initial EditableValue was empty but was not recognized as empty because the given DOM structure was wrong, so the toolbar ended up as the value. In the past this was not detected and led to false conditions, so the tests made false assumptions. This is fixed now and the initial test for testing faked API failures has been adjusted to work now.

Change-Id: Ib27359d2dc56f2b902332da7ce4cbc7bf389b22d