Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  2a7fe3d6713a07d42ccb4c3bbc1086c6af5d4935 null Fix QUnit test for EditableValue which assumed initially not in edit mode
This is a false assumption which was true in the past. Since we display empty input fields initially if the EditableValue is empty this test should have failed. This was not the case because of DOM changes which were not adjusted in the test file so there was a value set initially so the EditableValue never was in edit mode initially.

Change-Id: I0099c9a951cb87de53bef11c5a6baa02e9c0954d