Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  a9565769692857527e22da07418256b417c27e82 null Refactored some of the EditableValue tests.
Tests do not execute assertions in startup and tear down anymore now. Added a generic function to fabricate new EditableValue instances. Upon this change we could refactor tests for other EditableValue types to use this common base and always execute these basic tests instead of only testing functionality which is specific to the inherited EditableValue type (which is what we are doing right now).
There are still 17 failing tests, these will be fixed as a follow up.

Change-Id: Ia2791687772bdb64f6f9b8586b13855859104361