Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  a7017200a97079d6dd83c539bd6e4d10b0fa8940 null Adjust DOM structure of aliases and add edit button for non JS-version
For aliases, the required div.wb-gridhelper for styling is now part of the generated DOM already and won't be generated in JavaScript.
The 'edit' button will be placed as sibling of the '.wb-aliases div.wb-gridhelper' node. This is equivalent to the 'add' button for aliases in JavaScript mode. In JavaScript mode the 'edit' button is placed in '.wb-aliases .wb-ui-propertyedittool-editablevalue'. It should not be necessary to differentiate between 'add' and 'edit' button in non-JS though.

patch set 2: proper clean-up of DOM structure
patch set 3: rebase
patch set 4: fix regarding QUnit test

Change-Id: Ic063b6103104f187781ce376b4ba20ce8ba5bc5c