Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  a539ded24a65981a68bfe7c4166797e30a91883e null Fixed bug introduced in I6fb93b46, fix failing selenium test
This bug happened because the promise.promisor.apiAction always has a value since I6fb93b46. In the case we are  filtering for now, promise.promisor.apiAction was set to undefined before I6fb93b46. After changing the apiAction to always have a value promise.promisor.apiAction was no longer undefined so the condition was true even though this was not intended. This extends the condition as it should have been originally.

Change-Id: I3d115afee12d80c87bc869d43e645218bd9adc66