Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  845ae38e22a8453b6391e4aa61b131f048a4ffef null handle shell errors in runtests.sh
Whenever node has an error, the stderr is intercepted to a result file
and nothing is shown. That might confuse a newcomer. This change
intercept the node exit code and will show the content of the result
file for user to investigate.

A typical use case is a node module depency missing.

Change-Id: I946771aea4289e2a5389dbca8b928d83722b234a

shell function to echo to stderr

The simply warn() function would let you easily echo a message to
stderr. Examples usage:

 warn something got wrong
 warn "really I mean it"

Nicer than writing 'echo "message" 1>&2' isn't it ?

Change-Id: Ie85cc885b2753b572f6489ed215cd0eea4c70292