Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  ef81916d3da0f25c793aa4c53c25969b002a9c9b null Fixed failing QUnit tests after Id0cb4088 got merged early
The main problem was that if wb.ui.Toolbar.Label.stateChangeable is false, some of the state handling didn't work logically anymore. This has been fixed now. Another bug which was detected while fixing this is that empty toolbars always returned disabled as state. This has been changed to return enabled (which isn't much better, either the state should be remembered or we should re-thing the state system).

Change-Id: If6545d9b24a415c1d3da1cb72ba1e8dda568dff7