Updated mediawiki/extensions Project: mediawiki/extensions/Narayam  d63a9ba8bd5d8482ed3de9589b026a0448a6a611 null Remove the expensive event listeners
In every keydown, keypress, focus, blur on the document, Narayam
was triggering the jquery selectors - this is very expensive. This
was done to enable Narayam on future DOM elements. This is very rarey
used feature and its cost is high.

As of now, I am not aware of any usecase which use this. If we notice
such usecase we can add them in some UI context with narrow selector
scope.

Browsers like IE is very slow in executing jquery.on on selectors. See
Bug 40630. Eventhough it was solved by jquery upgrade, it is not good
to keep this feature in Narayam.

Change-Id: I7c382913179af913aafd7ecca4dbc18a988ec1f9