Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  56e79670a565cd9e6d3e3fd08af8919f07c1dfaf null (Bug 40381) Use prefixed id
Updated the code to both handle prefixed and non-prefixed ids.
The prefixed ids will be used for keys, but will not be used
in the "id" field as this is numeric.

Checking for a prefixed id is done in a try-catch, which is
somewhat awkward, but fixing this (that is creating a better
test) should be done in EntityFactory, and also in another
changeset.

Patchset 2: Removed unused boolean

Change-Id: Ib02c655485087c4322233a6e4f8f089bcf3c9874