Updated mediawiki/extensions Project: mediawiki/extensions/Vector  2d79ca4253b43aeba31447c1bfd4426970d9b62e null Decrease load time of footer cleanup module.
Module is now loaded at the top of the page to allow dom maniuplation
to occur immediately which prevents nodes from flashing.

Checking for advanced editor on window load as the advanced editor
loads asynchronously.  If found, remove edit help as intended and
trim the | symbol or anything outside of the cancel text span as it
is no longer relevant.

Removed editcopywarn styles as these are now in the en.wiki
mediawiki message.

Change-Id: Ia1d41ae24c869ce95b70a6c31dcd860fbd0e29b7