Updated mediawiki/extensions Project: mediawiki/extensions/MoodBar  38e2c44032519acfb95ec61c3ab85149346e8f40 null Stop moodbar from doubling in width. Revise hacky ie7 code block.
When clicking on 'what is this', the moodbar dialog would grow
over double in width.  Setting a width on the text container
prevents this.

The existing ie7 width manipulation is a complete hack.  Funny
enough, the show / hide part of this hack was outside of the block.
I moved this bit into the conditional in an attempt to make this
slightly better.  I will find a way to elminiate this.
For now, it is at least better than when I found it.

Change-Id: If1e2e4823dfbd956f4c9e398e57f4af1b3bba5e9