Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  aa29ac93fcd520faabc60885c2067c753666c352 null First attempt at round-trip testing on dumps
This patch adds a command-line utility for testing an article for
round-trippability. To come: A script that runs the script for a
random sample of titles in a dump.

Also, living up to my reputation, added a plain and JUnit XML
version for the output callback.

Bonus: Compatible with async.

Change-Id: I55c31a83ee8ed0fa0bc35426446ba36d50f60bff