Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  ea646dc8be1180b64c8d615cdd299e571be46aeb null enable dynamic sections
mf-toggle.js replaces toggle.js and requires jquery
serve page with empty sections, retrieve html for sections via javascript
note: currently no fallback for non-javascript users

load sections into DOM when clicking on the section headings

mark page with mw-mf-special if it's a main page or special page to avoid
ajax requests to api for these pages where it is unnecessary

hide section_anchors until expanded

Performance:
https://docs.google.com/spreadsheet/ccc?key=0ArbzKvV50qF6dEFzR0pCbUxfVy1ybGVPZU9yRTU0OUE

I ran 31 samples on the Barack Obama article with this commit and before this
commit. The results showed that time for js modules (search,references etc) to be enabled
decreased from an average of 1.98s to 0.20s -
a 89.9% increase. However toggling took longer to be enabled - 2.34s increased to 2.01s
- a 14.32% decrease in speed.

Provided we find a way to serve section content to users without javascript there is
no doubt we are improving their performance, but we must reflect on whether a 14.32%
decrease in time to toggle is more significant

Change-Id: I5571ba8c7e5a6a0b4160a2164c6eb2a84bddcd17