Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  cb30a3caecbcc2eb82ffbea04cd93ac3e6a9f4c2 null (Bug 41034) Refactor ApiSetItemTest to avoid timeout problems
There is a pesky timeout problem that could be due to the number
of tests inside a single function. It is now broken up in several
methods and each of them should finish in shorter time.

There are also some changes to ApiSetItem to avoid problems when
the requested objects (page, title, etc) can be non-objects.

Change-Id: Ie2a6eff6b6217e683d83b02f62977224183b0538