Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  82ed7d78f926597cd58d1f3d0ddf9c8570850363 null Extend compatibility of String and Blob
Strings and blobs will no longer be distinguished in future releases since SQLStore3 does not need to impose any length restrictions on string data anymore. However, to maintain compatibilty with SQLStore2, the distinction is preserved for now.

This change unifies string and blob handling, especially in datavalues. This fixes a problem with displaying stored Text/Code values when using Store3.

Change-Id: Ic3878b1570d41e7792ac065369758c4712e4e262