Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  4b8b1b86c22d47ae46fbadbc5674cbce06a404ca null Fewer database SELECTs for getting property values
Retrieve data more selectively when looking for a concrete property. So far, this was done by restricting to tables based on the property type, but this is no longer ery effective since there are many tables now. The new code will read at most one table for getting the values of one property, as it should be.

Change-Id: Ia386578e9acd402cc7f702d758c15e3ed3af9bc3