Updated mediawiki/extensions Project: mediawiki/extensions/PostEdit  68b8875891c2ae5bd7b4e8f377ded5e8c49731f7 null Set config var on post-edit state
OpenTask is currently borked because it's looking for
a wgPostEdit config var, which we had stopped setting. That
made me realize that other bits of code might be interested
in post-edit state too. Consequently, PostEdit should set
a global config var when it detects post-edit state. This is
accomplished in the following way:

	mw.config.set( 'wgPostEdit', true );

Change-Id: I9380b0377eb04e5781f8aac000938c9de744bcb5