Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  1ec3e7981dbeb1a881c38ed940c83e34d640d878 null Make ApiBotEditTest robust agnst non-sequential IDs.
ApiBotEditTest was assuming that IDs of consecutively created items
would be consecutive integers. This is not guaranteed, and frequently
not true in the presence of blacklisted IDs.

This change also fixes a problem with interpreting the recentchanges
feed returned by the API: the latest edit is not necessarily the first
rc entry returned by the API, because the order of edits that were done
in the same second is undefined.

Patchset 7: Did a rebase before submitting

Change-Id: I6ec7a388a116e56c5471d761d8a9611fa98b26a8