Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  6441a9a48ff161bec3ffd248c517617bf9d86e0f null Selective Serialization Episode IV: A New Hope
This is the first attempt at selective serialization, whose magical
arms shall soon embrace us and solve all intractable problems. The
end.

Your wait is over. It's here. Review it!

There are also tests! Pass in the --selser option to the normal
parserTests.js script to run them. They may not be totally
complete yet, but they're a good start.

Change-Id: I67b5d90298ae4f7a475021ebeb514cd742c4ee65