Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  a8a2d134a1cc585fd3a31e2968956e159a3588dc null Don't wrap sequence of <link> tags (categories etc) inside paragraphs
This prevents the cluster of categories at the end of a page from being
wrapped in a paragraph.

The behavior is nearly correct according to the new parser test. The test
should pass once the indent I threw in to spice things up is no longer
converted to a pre.

Change-Id: Idfdfda61362858ef663f39c70482b9147ea48767