Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  7d254203cf95a7aaa15841c4f1125f0040983c43 null Protect StubSemanticData agains DI exceptions
If the database contained data about a property whose name is no longer legal, an uncaught exception would be thrown when unstubbing all properties. The usual cause of this situation is that a special property got removed and is no longer known to SMW (because it vanished in a later release, or because it came from an extension now disabled).

Change-Id: Ic3d6d169045e800dbdd706f8963b99e62828a0d2