Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  46098b08389a37facff34064ea8f13cb9b60cc2d null (Bug 41667) remove onPageContentLanguage handler.
RepoHooks::onPageContentLanguage is no longer needed, because EntityHandler
overrides getPageViewLanguage() to return the user's language, and
isParserCacheSupported() to return false. There is no need to pretend that
the page's original content language is the user's language.

If this should cause problems, it is suggested to implement whatever logic
is desired by overriding EntityHandler::getPageLanguage directly.

Change-Id: Ia0ba651836d9151f7a58e89cfc10d49f8c8364a2