Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  437fc651d531b5645f5ef29f0f784147558d2994 null Actually report errors to the server
The clients would just skip merrily past an error, because the
serializer wouldn't pass them back, or even raise them....a giant
try/catch prevented it. Now everything appears to be forwarded
properly.

We also now have an env.errCB method that we should use to report
async errors in the parser pipeline--there's a try/catch hooked
up for the synchronous bits, but the rest of it should be added
as necessary (quoth gwicke).

Change-Id: I77c2e17471cc00d177f472fa7f7cf0ef03b1093f