Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  62eea42f0e058acd2ec79ce2432aa665f549c6bf null Special-case handling for template-generated list items.
If bullet has a content of '* bar',

* {{bullet}}
will result in
<ul><li>bar</li></ul>

and

* foo {{bullet}}
is expanded to
<ul><li>foo</li><li>bar</li></ul>

See https://bugzilla.wikimedia.org/show_bug.cgi?id=529 and Parser.php,
method braceSubstitution, line 3466 for the gory details how this came to
be.

The PHP parser also strips empty lists, which causes it to swallow the deep
prefix in

********* {{bullet}}

We don't currently do this in general, so perform a custom version of it
here.

With this patch, two more parserTests are passing, and pages like
[[1935 Philadelphia Phillies season]] round-trip without source diffs.

Change-Id: I2241900951b665422b459bb8e9684000cd7dbf4f