Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  9baaa0f417979832583c09d5f67724b1cfd9fbfe null Support variable SQL backends in SPARQL Store
SMWSparlStore used to be a subclass of SMWSQLStore2, which made it impossible to switch to SQLStore3 in a clean way. The change turns SMWSparqlStore into an adaptor for a base store object that can be any SMWStore. The class to use there can be chosen by setting SMWSparqlStore::$baseStoreClass; the default is SMWSQLStore3 now.

Patchset 2: add missing @since.
Patchset 3: fixed documentation

Change-Id: I7e60704e4faadb32aac1899d2047e8c66dbb83c3