Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  c443b58550caa13ae683de8bc5ec6f7c85817722 null Fix sortkey retrieval in Store3 (Bug 42033, ...)
Since sortkeys are not stored in a property table, but in the SMW IDs table, methods need special code to handle them. This was missing everywhere, but has a practical impact only for getPropertyValues(). Where unexpected, this caused warnings and errors (Bug 42033, Bug 42034) but also spurious problems with ordering of results (only in certain cases).

The change fixes this for getPropertyValues and adds remarks/todos in all cases where sortkeys are not currently supported (but not practically needed either).

Change-Id: I69b7900152efb72b3fc5e31f3b68a07b22a2e67d