Updated mediawiki/extensions Project: mediawiki/extensions/SemanticDrilldown  01cc9ab7aed22dddeda2b89932143529c08f83e8 null Support added for SMWSQLStore3
An extensive commit - everything related to querying the SMW database
tables was made more generic, so that it could support both SMWSQLStore2
and the new (in SMW 1.8) SMWSQLStore3. As part of that change, the
HTML and Javascript that handle arrow toggling were changed to use jQuery;
which in turn led to SD's two ResourceLoader modules being merged into one.

Change-Id: I3155f4d42fc7dff0f0d95179126ae5736a370614