Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  04a9cb457187137765d345dd21e6743e1242a75c null (Bug 39836) Auto-resolve edit conflicts.
This implements automatic resolution of merge conflicts.

Not that this makes it necessary to apply atomic edits always on top
of the designated base revision, so a valid patch can later be generated
against the base revision and replayed on top of the latest revsions
(in git terms, this is a rebase).

Change-Id: I344d76812649781c83814afb8e9386ff66fc9086