Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  fe1ffe439fc1e9b888ca76f2c86d42f8916b5ee2 null (bug 42063) linked data interface for wikibase.
Special:EntityData provides an easy way to retrieve the machine readable representation of an entity.
Canonical URIs could point to that page, so they resolve to the entity's json represenation.

This implementation supports the format parameter for all formats also supported by teh API. Furthermore,
the desired entity can be specified using the "id" parameter as well as using subpage syntax.

This is a baseline implementation with several TODOs left in the code. It is intended as a proof of
concept, lacking several features such as http cachign support or retrieval of specific revisions.

Change-Id: I4c9f7703004e63622eb63bec25d7d8f8d7e76c9f