Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  1df181018ff384847a9c30cce1950b298663795a null ClaimList now uses HashArray rather then HashableObjectStorage
In retrospect we currently actually do not need efficient by-guid lookup since we
can just get away with looping over all claims and then selecting the one with the
correct guid. Leaving it as it is for now since its quite possible we run into
a use case for it soonish. If we dont it can be changed back to use HashableObjectStorage
if we want to make it as simple as possible

Change-Id: Id7f0aa283e368007a4f88bee943db6f17a881a12