Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  51c78fc43d034f12aa25fdc92b97bf0f739d6939 null Move serializers from repo to lib
Already found out that they are not API specific (luckily we split it up in time :) since we need them in the entityview.
Now realized we will very likely have similar use for it on the client when we create intrefaces for editing factboxes
and the like, so moving this to lib.

Since it's clear this is not API specific, it needs some renaming, which will be made in a follow up

Change-Id: I1fdc8c28b8df0ea831f0689461693a68b6214597