Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  702f44d77323c9b335514b004edea88ca2602f52 null Fixed Bug 42032
The bug is caused by more than one entry referring to special properties in the SMW IDs table. There are multiple reasons why this has happened:
* Predefined properties are stored by their keys (e.g., _TYPE) but if they also have a property page, this page will be called like their current label (e.g., "has type"); this was not matched
* The refresh script carelessly created refresh jobs even for predefined properties, thus refreshing pages with titles like "TYPE" that never existed.

Both problems surfaced due to the new way of producing Special:Properties.

The new code requires Store3 to be reset and recreated to fix any existing problems caused by the earlier misbehaviour. A simple refresh is not going to delete the obsolete ID entries.

Note for review: SMW_Sql3SmwIds.php applies some inconsistent naming for new symbols. The old ones should be fixed in another commit, but this is not release critical and would require more far-reaching changes since some of the methods that should be renamed are public. Can be done in SMW 1.9.

Change set 2: Using property IDs (with property keys instead of titles) in all places where a property is used requires us to also reconstruct wikipage DIs from such IDs now. If a predefined property is used as a property value, then this is necessary. The DIHandler for wikipage has been augmented to do this, and all code that created SMWDIWikipage from SMW ID table contents is now calling the handler instead. Also fixed a lot of documentation.

Change-Id: Ifd6267cbd3fca1c3bee1b1927c70b91bd8128356