Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  1b9461c41d3e544bf77bcf92133514099c212adb null Migration Store2, Store3 via JobQueue
## Description
There is some db select overhead but it is a trade-off to keep the actual
update process within a job that can be put on hold or restarted without
locking down the database during the update process.

## PS 2
* Add option '--replace' which migrates and removes the data from Store2

## PS 4
Re-factored SMW_migrate.php where it now only initiates one batch job.
This batch job when executed will when in 'init' state select
the smw_ids and pack id's into a lot job of 500 and creates self-reference.

Of course you could sort before the db select and just give the
start and end position of the select but since I'm lazy here I just
do a simple serialize/unserialize of the available id's and pack them
into batches of 500.

## PS 5
Delete replace option but introduce --b option to choose the batch size
(by default is set to 500) and --i as sleep interval (default set to 0)
to allow to take away some databse insert pressure while an individual
batch is set to sleep for a given internal

## Example
... SMW_migration.php --migrate --i 5 --b 300 means to migrate store2
to store3 with a sleep interval of 5 during the execution of indivual
batches with the size of 300

In fear of locking down the db, use $wgJobRunRate [1] to control the job
execution.

[1] http://www.mediawiki.org/wiki/Manual:$wgJobRunRate

Change-Id: Ic4b53e6c7c73f3d28c7af75211253981eb9ea28b