Updated mediawiki/extensions Project: mediawiki/extensions/VisualEditor  c9d6d359683830f5f7f45209d889a7b1b236ab7d null (bug 42219) Fix for document sync issue
When working with a document containing only a slug in an empty paragraph, tree synchronization would break because it was trying to rebuild a non-existent text node.

This change makes the rebuild always occur on the outer range, rather than the inner range, which prevents absent text nodes from being asked to be rebuilt.

Thank you to Roan for debugging this for like 20 min.

Change-Id: I8c3dad921ace395f0694f77cec44305a680657fe