Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  068b15064a3e10c1c9893bc2dfebce72d8060c52 null mw.Api's getEntities parameters work with arrays now (broken before)
When passing arrays to certain parameters of getEntities(), this would have failed because the Array.prototype.join() was called on the arrays but the result was not assigned to anything.
Also moved the languages parameter to the front, behind the props parameter since it seems to be more important than the others which are more likely to be omitted.

Change-Id: I1e7efc4570668f6ff55a3c1763c2292225097b09