Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  0e5abd3cc9dbf9321943e8b7d64fe8e365d479c7 null snakview will now display a message if a property with unsupported type chosen
Before the commonsMedia type was chosen for all properties, this was hard coded in the snakview. Now we get the property ID by the entity selector widget and make an API call to get the data types of the property. This could be optimized if the search would actually return the data type as well.

Depends on I7a364cda in DataValues extension repo.

Change-Id: Ic728186703632be4f0a51dae9a0eff15fb2c8ded