Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  4c27b5deac16514e5eb3368e9b7b9cdccde26e22 null (Bug 41624) Add labels to user contributions
Some of the special pages sets the html used as link text
up front to full text version of the title and then the test
for the html will fail. By testing if the text is the full
text it is possible to try to make something nicer.

Patchset 2: Removed sporious code from another patchset

Change-Id: Iadf1b3deb695d3bf4d5670f9419b3ec3e4e15ebb