Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  659a588bcb46b7ea4b2b37ef78348ef0718bb2f2 null Add Banner Callback for CN Impression Reporting
The insert banner function has been restructured for my sanity
whilst also allowing the banner to modify the impression report
call. It does this by looking for a function declared by the
banner (mw.centralNotice.bannerData.alterImpressionData) which
may return True/False depending on if the banner declares that it
showed or not. It also has the option of modifying the metadata
that is passed along to the record impression call.

Change-Id: I4782a96007a974d4c156a2153c85b6285dae8595