Updated mediawiki/extensions Project: mediawiki/extensions/VisualEditor  d4b39100139c92feef087ec6a91fbc4e81b29488 null (bug 42277) First character in empty document behaves strangely in Firefox
This was because Firefox intially puts the selection before the first
paragraph, which translates to a model selection of (0,0). Typing with
the cursor at that position causes bad things to happen.

CE normally fixes up the selection when this happens, but it doesn't do
this automatically on initialization. So I added some code to ve.Surface
that causes this fixup to happen.

Change-Id: I0dcee3a29c1242c49ec30c743f1b69686fbb8436