Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  5d81db5136b168302872f9301083649fe2e1d9a4 null Collapse optional newlines in table syntax into a meta tag
This avoids the creation of (foster-parented) paragraphs for wikitext like

|-\n\n\n\n\n\n\n\n\n\n\n\n\n\n|foo

This avoids several screenful of whitespace at the top of Obama and other
pages without affecting round-tripping.

The approach taken here is a bit simplistic, as it won't take
template-generated newlines or leading newlines in template-generated table
rows into account. A more robust fix would probably be in the ParagraphWrapper
on the expanded token stream.

Surprisingly, six more html2html tests are passing. Round-tripping in pages
like Obama and parser tests also does not seem to be affected.

Change-Id: Ib300dc77fbe74eac893beda9684dbe67d7d5d0f9