Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  42312624f9f872216e7a3bb982931e5355857765 null Make Cite <references> group attribute handling more robust
The PHP preprocessor returns a references tag with an empty group attribute,
which we did not map to the null reference group. This caused the references
listing to vanish when using the PHP preprocessor.

This patch makes sure the attribute is converted to a string, stripped of
leading/trailing whitespace and finally set to null if no string remains.

No changes in parser tests, since Cite is in a separate extension.

Change-Id: Ib3de375225947a78c362370be8d78504fa24def2