Updated mediawiki/extensions Project: mediawiki/extensions/LabeledSectionTransclusion  d862f3229842eabd0f4912b996fd1f8a45439e25 null Default entry point + rename LabeledSectionTransclusion file
This is mostly used for continuous integration system. When testing an
extension we use an automatic loader in LocalSettings.php which iterates
over directory names under /extensions/, the loader then attempt to
require() a php file named the same.

The entry point loads both lst.php and lsth.php so we can run both
parser test suites.

The file LabeledSectionTransclusion.php have been rename to
LabeledSectionTransclusion.class.php in order to allow the use
of the conventional name.

Change-Id: Ia91f8b232fa85aa09c18964c535c269790e92186