Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  51fc0e5013c7cb065795184b2f14dbf4fc12eb35 null Using the Claim templates in the Claim widgets now, introduced $.claimlistview
The Claim widgets are now using the same templates which are used in the backend. Also, moved some code from the entityview into a claimlistview which is managing the list of claims now, putting claims into the same section of the list.

One important todo introduced by this is, that we update the local store of property information at the right time. Right now, when adding a new claim which has a Main Snak for a property which is not yet represented on the same site, the view can not display that properties name and will display some ugly, untranslated message instead.

Change-Id: I8ca1a4edc0ba1eb89bdc46083d828d4f7e801328