Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  adb41a02aacdf5249fd20c328248da9b4e0b1baf null (Bug 42526) Workaround for failing tests
This is a workaround for tests failing due to some other tests are
leaking changed state for $wgUser. A followup to this could be to
identify the tests that leaks "sysop" and "wbeditor". The last one
doesn't seem to be a problem, but the former is a problem during
tests where actions like "delete" is supposed to fail.

Patchset 2: Removed sporious stuff
Patchset 3: Additional crap
Patchset 4: Removed commented out stuff

Change-Id: Iaa705f7262a9382e70efc71df66a4a772153e0f6