Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  ff630209d89bb50614eaca035ea2b9d3f362e544 null Mobile replacement for Special:Watchlist, initial version.
So far includes a simplified overview display, bnd an initial drilldown into individual changes which is meant to be shared with history view etc.

A basic filter is available to select between 'all' pages, article pages, article talk pages, and other pages -- this needs further work as well.

May need to consider more issues of tagging/marking.

Note that some of SpecialMobileWatchlist's code is copied from core SpecialWatchlist, with the intention that it will be refactored further into a base that works with both desktop and mobile.

Change-Id: Iee1a4bfde4369df0f1f0b1dba3a65d45219e8efa