Updated mediawiki/extensions Project: mediawiki/extensions/ArticleFeedbackv5  9158f558020d440ba7fea0b900a35e48fbb9bb0d null Remove dependency on E3Experiments for clicktracking
E3Experiments had to be installed to use .stall functionality,
to ensure that en event would not be propagated before completing
the clicktracking call, which is async by default.
The dependency on E3Experiments is no longer desirable, as AFT
will need to run on wiki's that do not have E3Experiments enabled
(in this case: dewiki). Another (even safer) approach has now been
implemented to ensure that an event does not propagate before the
clicktracking call is completed: making the call syncronous!

Change-Id: I8da7da4b697f63df66fe5f8eb6866d6c49e636dc