Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  6079bacc21d0d7ec95e1c98c756e68a7249ef129 null (bug 42729) Don't load MW-1.21-only class on earlier MW
In MW 1.20, ApiQuery instantiates all generator modules in its
constructor. Therefore, this generator module, which depends on a core
class introduced only in MW 1.21, must not be used on earlier MW
versions.

Instead of a version check we just check whether the needed baseclass is
known to the autoloader. If it is, we may use ApiQueryUploadCampaigns,
otherwise we mustn't.

Change-Id: I314d5dd1ee3e6e5b95ad8bf888de229384ee1516