Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  bb979904262c8a5d465b73be689a98a8b460676c null (bug 43068) Display proper label for PVSnaks using wikibase-item properties
The label will only be shown when adding a new property. We still have to implement that this will also be shown when loading the page, after initializing the JavaScript.

Renamed wb.properties to wb.entities, this might contain some items from now on.

(!) Depends on Id5de9f6a in DataValues repo.

Change-Id: I3b0b09171d11c3704798f071b65b204e81439517