Updated mediawiki/extensions Project: mediawiki/extensions/E3Experiments  407460f41486215158db7701ba02599e9ff0508a null (bug 42982) transform the right part of the page
In account creation, if the server-reported error included a table
jQuery would operate on the wrong elements.  This fix:
1.  Avoids the specific JS error that halted the munging leaving the form
invisible.
2.  Is smarter about finding the right table to modify.
3.  Doesn't hide so much of the page during munging so if JS fails the
user sees something.

The last means there could be some flash of content in a slow
browser or on a slow connection, but I think it's worth it.

I considered catching all JS errors in modifyPage() and always restoring
the hidden form, but that seemed more invasive.

Change-Id: Ic76887e6df2caa5223fc9ba2c8e16b004a4bd805