Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaMessages  19fa5c880492ef299b20cb93f8a69e5842675ff0 null Add Wikidata license terms
wikidata.org uses a license footer that is significantly different
from the one on the other Wikimedia sites. This makes translating
it to other languages hard: unlike the default license footer, which
can be comfortably translated in translatewiki.net, the Wikidata footer
must be translated as a system message in wikidata.org

This change allows easy translation of the licensing of Wikidata
to all languages. It relies on the fact that $wgDBname for Wikidata
is wikidatawiki; if that is the DB name, then a different message
is used.

Change-Id: I6cddbdc65df34d3128f5b46274da36e13f84d206