Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  47639d86f3c79dedc0fc175229b33b7dac5fbf84 null (Bug 43202) Cleanup of claimview toolbar responsibilities
Toolbar save button will now only trigger stopEditing of Main Snak rather than controlling the API interaction. The toolbar will now listen to a new 'snakviewafterstopediting' which will retransform it to a non-edit toolbar.
This fixes that the enter button would not work when pressing enter to save a Claim (Bug 43202).

- patch set 2: updated TODOs

Change-Id: Ib5e6f7b2babc8807b7d754997140128cd6132b8b