Updated mediawiki/extensions Project: mediawiki/extensions/OpenStackManager  6a325317026e89dd76dd72e69574a6b13879bbfd null Now using real groups with real member lists for posixgroups for Projects.
ProjectGroups are posixgroups with manually maintained member lists.
Whenever a Project is created, deleted, or has members added or removed,
the same action will be called on the ProjectGroup.  This keeps the
Project and corresponding ProjectGroup member lists in sync.

There is also a new maintenance script 'syncProjectGroups.php', that
can be run to ensure that every Project's ProjectGroup member list
matches the Project's.

This allows applications that use LDAP to actually search for members
of specific posix groups.

Previously, ProjectGroups were maintained as virtual static groups,
pointing at the Project.

Change-Id: Idb33f2f48893a9afcfb4e008372386637c7a5bda