Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  1bd6a0c1bb911beb0719a1191bff70ba47d26888 null Handle link trails better
Fixed the bug where [[link]]s, if modified in the HTML to be [[lin]]s, would
round-trip to [[link]]s still. The behaviour is now reasonable.

Basically the difference is, we now check the actual href of a token instead
of only using the shadowed attribute values, and we don't try to cheat with
shortcuts based on data-parsoid elements.

No changes in parserTests, and VE will now be able to change linktrails (and
probably this will fix some other bugs as well).

Change-Id: I2a6cb59df691d7ccbcc7cc663249b5b696f2cae8