Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  65bffebb9e01c497839c41e22385690cf27379c6 null Upgrade to full EventLogging & declare schema dep.
This change removes the obsoleted implementation of EventLogging, and
adds a dependency on the new-and-improved EventLogging extension. The
schema dependency is declared but not used; I am deferring that to a
separate changeset.

The dependency is toggleable and turned off by default to keep the
MobileFrontend extension easy to install for third parties.

Change-Id: I4e1487f29561e84c78e8070a8f3f146c3e52b054