Updated mediawiki/extensions Project: mediawiki/extensions/TranslateSvg  c2fe7767d0a943c690791e4696fcce9cf665a2aa null Make the PNG thumbnail update live
A live-updated thumbnail allows users to perfect the formatting of
translated messages. Patch includes related fix to SVGMessageGroup.

There are essentially two parts to this commit, rolled together to
allow testing and to make it clear why certain design decisions
were taken.

The first is a series of hooks into Translate to amend the
API meta=messagegroups. The second is a JavaScript section that
uses the the amended API methods to provide the live update service.

For live test version, see
//translatesvg.wmflabs.org/wiki/Special:Translate

Change-Id: I9c75d38028adc7cbab48a1d17f97c30a0ae12123