Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  b93b63d51275bc7ba0f9b1f805c6d073f945589a null (bug 27316) Add option for (en|dis)abling multiple-file uploads
In addition to disallowing multiple file selection, we can now
disallow multiple files being uploaded in a single session through
a single configuration variable. This is used not only to disable
the creation of new file inputs and buttons, but also to disable
the multiple file select functionality if you've decided to leave
it on.

Change-Id: I694e7ab1f2dec461e781f6ebf672a66cd2ee12b5