Updated mediawiki/extensions Project: mediawiki/extensions/SemanticResultFormats  ab9a5a32d9ec8727c335f93a06363ded44b25b68 null Add blockUI/growl plugin
A plugin [1] that simulate synchronous behavior when using AJAX in order
to prevent user activity with part of the page until it is deactivated.

It also allows to send quick growl-like notifications to an user after
a "job" is done.

This relates to the forthcoming eventcalendar ajax update integration

blockUI plugin Version 2.54 (17-DEC-2012)

[1] http://malsup.com/jquery/block/

Change-Id: I81385b2a1624765fe657d8c3c518aca8b7763f83