Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  349681b0e1f686e7941500f1d6185d504fdbaf3d null (bug 42064) Call EditFilterMergedContent from EditEntity.
This is a partial fix to bug 42064, triggering EditFilterMergedContent from
EditEntity, so AbuseFilter and the like have a chance to filter edits
performed on Wikibase entities.

To do this, an IContextSource needs to be provided to the EditEntity instance.

This change requires the core change I99a19c93	to work correctly.

Change-Id: I5f5b4200f8a4db0e34790ad0c704ec9e17aff1f1