Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  9a57921a95a44f244751901b615c23be6272214b null Mark auto-inserted td-tags in the tokenizer so they are not RT-ed.
The tr- production inserts an implicit td-tag when content follows
the tr-tag on the next line.  This introduces rt-errors.  By marking
the td-tag with a 'autoInsertedStart' and 'autoInsertedEnd' flags,
the serializer will ignore this.

Fixes RTing of wikitext:

{|
|-
foo
|}

which in turn eliminates the zillion semantic errors on:
en:List of United States micropolitan statistical areas by population

Change-Id: I56645bd1a3ec1bb6af51d30524c9482ed1bc8a9b