Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  80b9bc371e1b14bd847ca0e9f972f579b5792f2d null (bug 43304) Moved snak type selector for snakview widget into its own widget
There is a jquery.wikibase.snakview.TypeSelector widget now. This one is only used by the
snakview widget internally since it is not really a widget displaying all possible Snak types
but rather those, which can be chosen by the snakview widget. Snak types unsupported by the
snakview are not displayed in the selector. This makes the widget rather useless for the outside
world and therefore there is no jQuery.widget.bridge used to make the widget available to the
outside.

* Also brings some minor changes to the type selectors architecture and handling.
* Less event registrations are required now since there is only one listener for click events
  on the menu now.
* using 'active' instead of 'disabled' ui state for chosen elements (even though it looks kind
  of stupid, styling might be in order)
* when resizing the window, the menu will vanish now - this might be changed again but it seems
  less tricky in some edge cases for now

- patch set 2: Renamed to SnakTypeSelector, updated some naming
- patch set 3: Renamed setMenuPosition() to repositionMenu()
- patch set 4: Updated destroy() method

Change-Id: I49a21e615a047a40721bbfe36abfbe35c1e66631