Updated mediawiki/extensions Project: mediawiki/extensions/TimedMediaHandler  7ac46b77d8e5a3ffdb863e4515e7c9a2452de9bf null Remove width rule for the "copy code" button
Setting a hard-coded width makes the button too small for
the label in many languages. Try it with Greek (el) and
Tamil (ta), for an extreme case.

If the width is not set explicitly, it is set automatically,
which should be just fine for this case, because this
button doesn't have to fit in with anything else.

Change-Id: I8921b373b59fd87d33f14a28d376be4eb2134b96