Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  91ff9bcdb53ac6607f776f35e12d4770dea7ef6d null Classify leading tab vs. space as syntactic diff
This reclassifies RT issues on wikitext like "\ta" -- found on at least
a couple pages.
- en:Telecommunications in Sudan
- en:Coorg National Council

We rely on selective serialization to hide these kinds of diffs normally, so
classify them as syntactic in our rt tests. Round-tripping these diffs without
selser would be a bit tedious. A partial rt patch was developed in
https://gerrit.wikimedia.org/r/#/c/43489/.

Change-Id: Ibab03b4739bb652e7e8fb9f9f09c5a749531b060