Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  f03ffdbf012c72454931964ad1a7002e643deb33 null Remove flash of unstyled content (bug 43739)
This refactors the way we serve javascript and stylesheets to
special pages. Special page modules are dynamically created in MobileFrontend.php
- each special page module as a scripts and css component. The scripts are loaded
at the bottom and the css at the beginning. The modules are created based on a naming
convention ( <lowercase special page name>\.(css|js) )

In process adjust filepage css/js so that css loads in head

Change-Id: I891f9b4fc1f4d06db0ddad828aba6a50efdf182d