Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  47fb8a406d1053453e248108c801cd7841735041 null Add support for attributes in table captions
Table captions did not actually support attributes so far, which caused
wikitext like this to parse the attribute as text content:

{|
|+ style="color: red" | foo
|}

This patch fixes this and adds better DSR to the caption tag. Bug 43582 should
be fixed too.

With https://gerrit.wikimedia.org/r/#/c/44167/, the following additional tests
are green:

+3 wt2html
+1 wt2wt
+3 html2html
+1 selser (though no change assignments for new tests yet)

Change-Id: I69704cefe31e4c2f720c7f1a2bf4a9b5a9b6835f