Updated mediawiki/extensions Project: mediawiki/extensions/EventLogging  ba531e0ea15fa62fef3c068c619ac1b34c03cae3 null add a _meta._timestamp field to server-side events
Server log files don't seem to get a timestamp when transferred over
UDP.
This change calls wfTimestamp( TS_UNIX, 0 ) to add a
"_timestamp":1358318635 to the _meta fields that efLogServerSideEvent
supplies. This matches the format of _meta.timestamp in
client-side-events-json.log .  For comparison, ClickTracking's
trackEvent() calls wfTimestampNow() to add 20130110065450 to log lines.

Change-Id: Id8310347bd9c9df8c1f3a40f10a93d6d261e67e3