Updated mediawiki/extensions Project: mediawiki/extensions/Translate  0a68c849e7eba68ce5fc87e699148d0219f89f3e null Adjust z-index for action bar to avoid overlapping with message editor
The z-index has been added since the text area of the message editor
was displaying on top of the action bar while the rest of the editor
was displayed below (in terms of depth). As a result, depth was inconsistent.
with this change, the action bar is displayed on top of the editor when floating.

Change-Id: I91a6e8ed62fa3f7fb55a2a389822d77480a84d65