Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  4592c00f4e2d26a1b8a0f827bc60e8a5d8e6b3c6 null Returns mobile special page module loading to wgResourceModules
This is in essence a partial revert of Change-Id: I891f9b4fc1f4d06db0ddad828aba6a50efdf182d,
which works fine in Beta when we have full resource support, but is busted
in production without ResourceLoader support, but otherwise maintains
the dynamic generation of mobile special page resource modules.

We can go back to doing this in the invocation of the
ResourceLoaderRegisterModules hook after we have RL support
everywhere.

Change-Id: Ie1e088732a11b3fd101a96d2e1959bf7598a35d6