Updated mediawiki/extensions Project: mediawiki/extensions/LiquidThreads  341e7e7476fffe1c186eab26775203c5be52933a null JSHint: raise max errors to 10000
We have lot of JavaScript errors in this extension so raise the
maxerr to 10000 so we can see all of them.

Change-Id: I478a8a3713ad5ce7d7dd515a034ab6d07278caca

JSHint: ignore jquery/jquery.autogrow.js 3rd party

We do not want to report JavaScript errors that comes from third party
libraries such as jquery/jquery.autogrow.js. This patch adds it to
JSHint ignore list.

Change-Id: I93d583d7095d62036c75428a9c7ff7dd277c70aa