Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  63abc1b1bfda62be958bf3d30afae77d98613c65 null fixed resultprintertest::instanceProvider
Change-Id: I6f866c932df910e7befaf4329d7e9c67686e38dc

Updated min Validator version requirement

Change-Id: Ie862bf84a94b58ee3f66e0a8fc881035ab0ce1b9

Fix test to not fail due to core returning ResultWrapper rather then true when using SQLite

Change-Id: Ie8b0677026e1184b2b9a607714cc938c49c00ca0

bug 42324: fix sqlite support in sqlstore3

Change-Id: I16336fd2a3425cbabda5808793e22cae25cfb729

Removed the need to call enableSemantics

The vast mayority of people dont care about this thing anyway.
And for those that do, tis likely the default is acceptable.
And if that's not the case, they can still set the global.
Calling the method as before will continue to work for now.

Change-Id: I6c645d170373a6a3dea829762a9636df0882a8cd