Updated mediawiki/extensions Project: mediawiki/extensions/PostEdit  4bf3b45db0f8100ed210c9c8e171de383fe45194 null (Bug 41805) Show PostEdit on edited page only.
Instead of a boolean, 'isPostEdit', track the ID of the page that was edited,
and set that ID as the value of the PostEdit cookie. In JavaScript, only show
PostEdit if the value of the cookie is the same as wgArticleId, but clear the
cookie regardless.

This change also binds the post-edit handler to PageContentSaveComplete rather
than ArticleSaveComplete, as the latter has been deprecated.

Change-Id: I67d3f2d07bd983e412c23a7dc65cdeb2a9f52c2e