Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  be568dbd732d218327632f7701466072a56db4e5 null (minor) moved jQuery.TemplatedWidget template related fields into options
the template options have been fields in the prototype before. It makes kind of sense
to make them available in the options though. An issue with having 'templateShortCuts'
as a field was that inheritance would not be possible with the jQuery Widget system
since only options will get a true copy by $.widget while other objects will be
modified on the base prototype. Our workaround for this only worked for one level of
inheritance (doing the copy manually in the prototype's constructor, can't define the
constructor of the new prototype created by jQuery.widget() though).

Change-Id: I18a62c2bf1bfd4c92ba53fd7775948009d16b130