Updated mediawiki/extensions Project: mediawiki/extensions/AbuseFilter  8cb215bc014d28ec7c29456bcac8a55763a86ef8 null Allow special pages as context for abuse filter.
The title in the context provided for an edit may be a special page
in some cases, e.g. when an extension uses a special page for creating
new content, and the final name of the content is not yet known
when the EditFilterMergedContent hook is called.

This change will allow special pages to be used in this context.

Change-Id: I5551bff69e3b617a437bd043fdc893dbdcd9f09f