Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  300d81981639116e2b802c2ad2f42c1eb3e72b0d null (bug 43332) Use the WikiConfig for magic words
This patch introduces a more general solution to detecting and round-tripping
magic words. Now we use aliases from whatever wiki we're using, and there is a
large list of default magic words to be found in the WikiConfig prototype.

In short, international wikis will now be able to parse and round-trip their
magic words, and even English wikis will be better off because we can now
round-trip __toc__, __ToC__, and __tOc__ instead of just __TOC__.

ParserTests: No change.

Change-Id: I61ba07334215e71695f44670f54c72f57d15925b