Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  de84c376c5770c91c4964e219f1558f656ed1787 null (bug 43909) (mingle 377) Include article to watch information in login form
If a user is coming to the login form from tapping the 'add to
watchlist' icon on an article, this will include the article to
be watched in a hidden field in the login form. This will later
be used to actually watch that specific article after login is
complete.

Change-Id: I8b7263557761610ffd87f8793d01775b5c85610b

(bug 43909) (mingle 377) Remove local storage component of non-logged in watchlist star tap

Change-Id: Ic1e9e8521240aeb0188ba5e650957fcd0befab24