Updated mediawiki/extensions Project: mediawiki/extensions/Renameuser  4f46caa5ba337406b69378ce92a1ec7e96e142ad null Add flag to RenameuserSQL to skip username check
It would be very helpful to use RenameuserSQL from CentralAuth to
rename CentralAuth users on the local wikis. However, CentralAuth
will rename the user in the local wiki's user tables (all wikis at
once), so it would be great if RenameuserSQL would not fail if that
update has already been made.

Change-Id: I5a8abd6fb2c1abc54c727adaffa4566d4f20617f