Updated mediawiki/extensions Project: mediawiki/extensions/Translate  c169668446607b6ad1aa69510ca958b5e0853b00 null Update the group description upon a group change
The group is initially set in PHP. Before this change
it was not updated when a new group was selected, and
the description of the old group was shown.

Now the group description is updated. The changeGroup()
function is changed so now it receives a message object,
which includes the id and the description. The description
is then parsed and updated in the DOM.

The existing calls to changeGroup() are changed
to pass a message object instead of a string.

Change-Id: I13da35cf20d7280e55e327b6a9928df387e80565