Updated mediawiki/extensions Project: mediawiki/extensions/EventLogging  1c7ae7889934fc237147ec2bb9773a43da32c303 null Relax conditions for inclusion of schema CSS.
If you want to reliably detect article previewing, 'wpPreview' isn't
enough: you need to also check for 'wpLivePreview'. Or use
ParserOptions->getIsPreview(). But I think it's simpler to just enable
the CSS for all pages associated with the schema namespace, the latter
being such a small sliver of traffic anyway.

Change-Id: Ibcda7377bed30938d5f2fb7257ecf335c019efa0