Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  13e48dc17622ea1bee7b6b42caa8110cbeecfb14 null Make the pages subclassing from SpecialWikibasePage respect i18n
Before this change the special pages would use a special local
system message for the summary. This creates trouble during
localization. So to solve this it is easier to simply always
create the message, thereby it would be possible to use the tools
in Translatewiki.net.

There should be a similar message for the special page CreateQuery,
but the name of the pages should be harminized. They should not be
called both "Create" and "New".

Patchset 3: Rebased and a minor fix of the two messages.

Change-Id: Ia4eee2658407fb76089cbf7e168cfe404e134277