Updated mediawiki/extensions Project: mediawiki/extensions/Scribunto  04a0a580e37a378ebfaae2a6010c09aa031b0d50
Add mw.title library

Note that fetching any title besides the one for the current page is
considered "expensive". It also records the title fetched in the
ParserOutput so it will be listed in pagelinks, just like #ifexists.

This also moves the ToString test formatter into TestFramework.lua, so
TitleLibraryTests.lua can use it too.

Change-Id: I799f3289a37fe1349b6bca5758829acf82cb718f