Updated mediawiki/extensions Project: mediawiki/extensions/OpenID  1a04304ba185394b6c7dbdef255120c60ea61994 null v1.005 breaking change: $wgTrustRoot to $wgOpenIDTrustRoot
Version bumped to 1.005, preparing for a move to 2.x.

This fixes a variable name which was not in conformity with the
other variable names. All global variables of the OpenID extension
start now with "$wgOpenID" .

This change is documented in a new file CHANGES, and further updates
to the Manual page will be done.

The fix prepares for a major change to version 2.x of the extension.

Change-Id: I7737a0cb3261053de0b65b716d55f89e4315d834