Updated mediawiki/extensions Project: mediawiki/extensions/OpenID  3aaa1f216a6b96b91ff08c8908100cd8bba66877 null (bug 44822) add verbosity if user tries to add already assigned OpenID
When user tried to add an OpenID which is already assigned to their account
or to another user's account, then the error text shows this conflicting OpenID
as actually given and  sent by the OpenID Provider.

Two message texts have changed their name and added a parameter $1 (the conflicting
OpenID).

Change-Id: I7d7b389a495c62b14656823e8976a3bbb32c9c37