Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  f6f2072bc6a3ca61b22179b78141fbbc2cf22965
Make client robust against unknown sites.

If the client wiki received links from the repo that contain
references to wiki sites that are not known on the client wiki
(due to differences in the sites table), this should trigger a
warning, but not a fatal error.

LIGHTLY TESTED. The change handling codepath needs another
look (ClientChangeHandler).

Change-Id: Ie21744510af795022b4c585d4e668a7662181b50