Updated mediawiki/extensions Project: mediawiki/extensions/EventLogging  2c7b929f02f3f1f65d8384d551ac772a49a85d82
Key formatter interpolates $wgEventLoggingDBname

Change Icfba05ba4b90c4241d18c33fd35a118554830182 sets $wgEventLoggingDBname on
test2wiki to test2wiki, making it a kind of metawiki-that-isn't-metawiki with
respect to EventLogging. If we don't interpolate the $wgEventLoggingDBname, we
risk collisions. We could interpolate $wgDBname instead, but that would
fragment the cache unnecessarily.

Change-Id: Id7aaad7e43a88af0192835bba58af3b0f6780ad4