Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  468fb0c41592b104a5328563594d18bbc93a9245
(bug 44744) copyright warning now using its own message instead of EditPage::getCopyrightWarning

Since the 'EditPageCopyrightWarning' is not suitable (without adjustments) to cover the needs of
the wikidata short copyright message, we introduce our own message in wikibase for this.

The following will be added to Wikidata's configuration to have the message displayed according
to Wikidata's copyright and terms of use.
- $wgRightsUrl = 'http://creativecommons.org/publicdomain/zero/1.0/';
- $wgRightsText = 'CC0 public domain dedication';

Without these configuration variables defined, the message will not look nice.

- patch set 2: Corrected spelling errors

Change-Id: I0a8328fbb50a8209fe9799f692898eccca378840