Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  94ec1e81601853e47974ab2d9bfafe689247d12b
Replace jsdom with domino, removing dependency on the native contextify module.

Domino is also reported to be faster, as it doesn't contain all the sandboxing
required to evaluate <script> tags in document context (which we're not using
anyway).

Use domino for all html5 parsing.

Update DOM version check for domino.

Change-Id: Iebf8fc5a20c3e64b4a0a36380525ca85c00c63e0