Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  8535068bf2f7971fad2d55bedbd139a9b577b8c5
(bug 41163, 44165, 44166) autocomments for claims.

This adds autocomments when claims get added or removed, or
their main snak is changed.

Caveat: When removing or changing a claim, the claim ID (GUID) is shown in the
summary. That is not very helpful. To changes this, some refactoring is
needed - this is already in the pipeline.

Change-Id: I0fd1ea5f98d3ee814f11ad57ece6911e1bdc2ff2