Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  1609a709f279956bd3ab1fddcc609527f9f98bbe
Update normalizeOut() to handle <meta> the same as <meta /> and <meta >.

The regexp used to normalize <meta> and <link> tags forced there to be
a space after the tag name... which there might not be after we normalize
out all the attributes of the meta tag.

(E.g. domino's HTML serializer doesn't add a space before the close tag.)

Change-Id: Id01a70dca569772c11f53e88043248db1ab646d5