Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  86578157fa616e35aa754edb8071e645e498ae9a
improvement for options handling of wb.ui.Base

We now make sure we overwrite the _options field, so instances will not share the prototype's
object!

As a visible effect, this fixes the language terms table's bug where all changes would be
saved to the last language in the list even though they were made for another language in the
table. This bug appeared because the options were shared between all Label/DescriptionEditTools,
this gets fixed by this change set.

Change-Id: Ibd4d327ccc69bfd75eb847f7528b73b4186b8fbb