Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  7d3cb0fe215985440531b724ea8f2695d3cd8804
(Bug 41882) Followup for restrictions for special pages

This adds restrictions for special pages that creates items and properties,
and for special pages that changes labels, descriptions and aliases.
It is also a minor change to an API-module for aliases to use only one
permission as it is only possible to use one restriction in the special
page.

There should be better tests for the restrictions, not only that the
pages passes if the permissions are set for the user.

Change-Id: I137187981ff4cc12864717583d0af6129aebfbc3