Updated mediawiki/extensions Project: mediawiki/extensions/Scribunto  b6ea64f1325433c574b39f7f360f7f5f4ec0fdde
Make mw.getCurrentFrame() work in console, add frame:newChild()

It would be helpful for debugging if a frame object could be gotten in
the console. To that end, add an empty frame when running in the console
and allow it to be returned by mw.getCurrentFrame().

It would also be helpful to be able to create frames with arbitrary
arguments, again for testing. Fortunately support for creating child
frames with arbitrary arguments already exists in core, so we can just
use it.

And for good measure, be sure to restore the $engine->currentFrames
array even if the Lua code throws an exception.

Change-Id: I1dc8602d63af75424f267c42a3743fabbc1827f7